Move TrStmt and ExpressionTranslator to separate files
