+++ Source/Dafny/Verifier/Translator.cs
@@ -10620,899 +10620,6 @@ namespace Microsoft.Dafny {
       }
     }
 
-    void TrStmt(Statement stmt, BoogieStmtListBuilder builder, List<Variable> locals, ExpressionTranslator etran)
-    {
-      Contract.Requires(stmt != null);
-      Contract.Requires(builder != null);
-      Contract.Requires(locals != null);
-      Contract.Requires(etran != null);
-      Contract.Requires(codeContext != null && predef != null);
-      Contract.Ensures(fuelContext == Contract.OldValue(fuelContext));
-
-      stmtContext = StmtType.NONE;
-      adjustFuelForExists = true;  // fuel for exists might need to be adjusted based on whether it's in an assert or assume stmt.
-      if (stmt is PredicateStmt) {
-        var stmtBuilder = new BoogieStmtListBuilder(this);
-        string errorMessage = CustomErrorMessage(stmt.Attributes);
-        this.fuelContext = FuelSetting.ExpandFuelContext(stmt.Attributes, stmt.Tok, this.fuelContext, this.reporter);
-        var defineFuel = DefineFuelConstant(stmt.Tok, stmt.Attributes, stmtBuilder, etran);
-        var b = defineFuel ? stmtBuilder : builder;
-        if (stmt is AssertStmt || DafnyOptions.O.DisallowSoundnessCheating) {
-          stmtContext = StmtType.ASSERT;
-          AddComment(b, stmt, "assert statement");
-          PredicateStmt s = (PredicateStmt)stmt;
-          TrStmt_CheckWellformed(s.Expr, b, locals, etran, false);
-          IToken enclosingToken = null;
-          if (Attributes.Contains(stmt.Attributes, "_prependAssertToken")) {
-            enclosingToken = stmt.Tok;
-          }
-          BoogieStmtListBuilder proofBuilder = null;
-          var assertStmt = stmt as AssertStmt;
-          if (assertStmt != null && assertStmt.Proof != null) {
-            proofBuilder = new BoogieStmtListBuilder(this);
-            AddComment(proofBuilder, stmt, "assert statement proof");
-            CurrentIdGenerator.Push();
-            TrStmt(((AssertStmt)stmt).Proof, proofBuilder, locals, etran);
-            CurrentIdGenerator.Pop();
-          } else if (assertStmt != null && assertStmt.Label != null) {
-            proofBuilder = new BoogieStmtListBuilder(this);
-            AddComment(proofBuilder, stmt, "assert statement proof");
-          }
-          bool splitHappened;
-          var ss = TrSplitExpr(s.Expr, etran, true, out splitHappened);
-          if (!splitHappened) {
-            var tok = enclosingToken == null ? s.Expr.tok : new NestedToken(enclosingToken, s.Expr.tok);
-            (proofBuilder ?? b).Add(Assert(tok, etran.TrExpr(s.Expr), errorMessage ?? "assertion violation", stmt.Tok, etran.TrAttributes(stmt.Attributes, null)));
-          } else {
-            foreach (var split in ss) {
-              if (split.IsChecked) {
-                var tok = enclosingToken == null ? split.E.tok : new NestedToken(enclosingToken, split.E.tok);
-                (proofBuilder ?? b).Add(AssertNS(tok, split.E, errorMessage ?? "assertion violation", stmt.Tok, etran.TrAttributes(stmt.Attributes, null)));  // attributes go on every split
-              }
-            }
-          }
-          if (proofBuilder != null) {
-            PathAsideBlock(stmt.Tok, proofBuilder, b);
-          }
-          stmtContext = StmtType.NONE; // done with translating assert stmt
-          if (splitHappened || proofBuilder != null) {
-            if (assertStmt != null && assertStmt.Label != null) {
-              // make copies of the variables used in the assertion
-              var name = "$Heap_at_" + assertStmt.Label.AssignUniqueId(CurrentIdGenerator);
-              var heapAt = new Bpl.LocalVariable(stmt.Tok, new Bpl.TypedIdent(stmt.Tok, name, predef.HeapType));
-              locals.Add(heapAt);
-              var h = new Bpl.IdentifierExpr(stmt.Tok, heapAt);
-              b.Add(Bpl.Cmd.SimpleAssign(stmt.Tok, h, etran.HeapExpr));
-              var substMap = new Dictionary<IVariable, Expression>();
-              foreach (var v in ComputeFreeVariables(assertStmt.Expr)) {
-                if (v is LocalVariable) {
-                  var vcopy = new LocalVariable(stmt.Tok, stmt.Tok, string.Format("##{0}#{1}", name, v.Name), v.Type, v.IsGhost);
-                  vcopy.type = vcopy.OptionalType;  // resolve local here
-                  IdentifierExpr ie = new IdentifierExpr(vcopy.Tok, vcopy.AssignUniqueName(currentDeclaration.IdGenerator));
-                  ie.Var = vcopy; ie.Type = ie.Var.Type;  // resolve ie here
-                  substMap.Add(v, ie);
-                  locals.Add(new Bpl.LocalVariable(vcopy.Tok, new Bpl.TypedIdent(vcopy.Tok, vcopy.AssignUniqueName(currentDeclaration.IdGenerator), TrType(vcopy.Type))));
-                  b.Add(Bpl.Cmd.SimpleAssign(stmt.Tok, TrVar(stmt.Tok, vcopy), TrVar(stmt.Tok, v)));
-                }
-              }
-              var exprToBeRevealed = Substitute(assertStmt.Expr, null, substMap);
-              var etr = new ExpressionTranslator(etran, h);
-              assertStmt.Label.E = etr.TrExpr(exprToBeRevealed);
-            } else if (!defineFuel) {
-              // Adding the assume stmt, resetting the stmtContext
-              stmtContext = StmtType.ASSUME;
-              adjustFuelForExists = true;
-              b.Add(TrAssumeCmd(stmt.Tok, etran.TrExpr(s.Expr)));
-              stmtContext = StmtType.NONE;
-            }
-          }
-          if (defineFuel) {
-            var ifCmd = new Bpl.IfCmd(s.Tok, null, b.Collect(s.Tok), null, null);  // BUGBUG: shouldn't this first append "assume false" to "b"? (use PathAsideBlock to do this)  --KRML
-            builder.Add(ifCmd);
-            // Adding the assume stmt, resetting the stmtContext
-            stmtContext = StmtType.ASSUME;
-            adjustFuelForExists = true;
-            builder.Add(TrAssumeCmd(stmt.Tok, etran.TrExpr(s.Expr)));
-            stmtContext = StmtType.NONE;
-          }
-        } else if (stmt is ExpectStmt) {
-          AddComment(builder, stmt, "expect statement");
-          ExpectStmt s = (ExpectStmt)stmt;
-          stmtContext = StmtType.ASSUME;
-          TrStmt_CheckWellformed(s.Expr, builder, locals, etran, false);
-
-          // Need to check the message is well-formed, assuming the expected expression
-          // does NOT hold:
-          //
-          // if Not(TrExpr[[ s.Expr ]]) {
-          //  CheckWellformed[[ s.Message ]]
-          //  assume false;
-          // }
-          BoogieStmtListBuilder thnBuilder = new BoogieStmtListBuilder(this);
-          TrStmt_CheckWellformed(s.Message, thnBuilder, locals, etran, false);
-          thnBuilder.Add(TrAssumeCmd(stmt.Tok, new Bpl.LiteralExpr(stmt.Tok, false), etran.TrAttributes(stmt.Attributes, null)));
-          Bpl.StmtList thn = thnBuilder.Collect(s.Tok);
-          builder.Add(new Bpl.IfCmd(stmt.Tok, Bpl.Expr.Not(etran.TrExpr(s.Expr)), thn, null, null));
-
-          stmtContext = StmtType.NONE;  // done with translating expect stmt.
-        } else if (stmt is AssumeStmt) {
-          AddComment(builder, stmt, "assume statement");
-          AssumeStmt s = (AssumeStmt)stmt;
-          stmtContext = StmtType.ASSUME;
-          TrStmt_CheckWellformed(s.Expr, builder, locals, etran, false);
-          builder.Add(TrAssumeCmd(stmt.Tok, etran.TrExpr(s.Expr), etran.TrAttributes(stmt.Attributes, null)));
-          stmtContext = StmtType.NONE;  // done with translating assume stmt.
-        }
-        this.fuelContext = FuelSetting.PopFuelContext();
-      } else if (stmt is PrintStmt) {
-        AddComment(builder, stmt, "print statement");
-        PrintStmt s = (PrintStmt)stmt;
-        foreach (var arg in s.Args) {
-          TrStmt_CheckWellformed(arg, builder, locals, etran, false);
-        }
-
-      } else if (stmt is RevealStmt) {
-        AddComment(builder, stmt, "reveal statement");
-        var s = (RevealStmt)stmt;
-        foreach (var la in s.LabeledAsserts) {
-          Contract.Assert(la.E != null);  // this should have been filled in by now
-          builder.Add(new Bpl.AssumeCmd(s.Tok, la.E));
-        }
-        foreach (var resolved in s.ResolvedStatements) {
-          TrStmt(resolved, builder, locals, etran);
-        }
-
-      } else if (stmt is BreakStmt) {
-        AddComment(builder, stmt, "break statement");
-        var s = (BreakStmt)stmt;
-        builder.Add(new GotoCmd(s.Tok, new List<String> { "after_" + s.TargetStmt.Labels.Data.AssignUniqueId(CurrentIdGenerator) }));
-      } else if (stmt is ReturnStmt) {
-        var s = (ReturnStmt)stmt;
-        AddComment(builder, stmt, "return statement");
-        if (s.ReverifyPost) {
-          // $_reverifyPost := true;
-          builder.Add(Bpl.Cmd.SimpleAssign(s.Tok, new Bpl.IdentifierExpr(s.Tok, "$_reverifyPost", Bpl.Type.Bool), Bpl.Expr.True));
-        }
-        if (s.hiddenUpdate != null) {
-          TrStmt(s.hiddenUpdate, builder, locals, etran);
-        }
-        if (codeContext is IMethodCodeContext) {
-          var method = (IMethodCodeContext)codeContext;
-          method.Outs.Iter(p => CheckDefiniteAssignmentReturn(stmt.Tok, p, builder));
-        }
-        builder.Add(new Bpl.ReturnCmd(stmt.Tok));
-      } else if (stmt is YieldStmt) {
-        var s = (YieldStmt)stmt;
-        AddComment(builder, s, "yield statement");
-        Contract.Assert(codeContext is IteratorDecl);
-        var iter = (IteratorDecl)codeContext;
-        // if the yield statement has arguments, do them first
-        if (s.hiddenUpdate != null) {
-          TrStmt(s.hiddenUpdate, builder, locals, etran);
-        }
-        // this.ys := this.ys + [this.y];
-        var th = new ThisExpr(iter);
-        Contract.Assert(iter.OutsFields.Count == iter.OutsHistoryFields.Count);
-        for (int i = 0; i < iter.OutsFields.Count; i++) {
-          var y = iter.OutsFields[i];
-          var dafnyY = new MemberSelectExpr(s.Tok, th, y);
-          var ys = iter.OutsHistoryFields[i];
-          var dafnyYs = new MemberSelectExpr(s.Tok, th, ys);
-          var dafnySingletonY = new SeqDisplayExpr(s.Tok, new List<Expression>() { dafnyY });
-          dafnySingletonY.Type = ys.Type;  // resolve here
-          var rhs = new BinaryExpr(s.Tok, BinaryExpr.Opcode.Add, dafnyYs, dafnySingletonY);
-          rhs.ResolvedOp = BinaryExpr.ResolvedOpcode.Concat;
-          rhs.Type = ys.Type;  // resolve here
-          var cmd = Bpl.Cmd.SimpleAssign(s.Tok, (Bpl.IdentifierExpr/*TODO: this cast is rather dubious*/)etran.HeapExpr,
-            ExpressionTranslator.UpdateHeap(s.Tok, etran.HeapExpr, etran.TrExpr(th), new Bpl.IdentifierExpr(s.Tok, GetField(ys)), etran.TrExpr(rhs)));
-          builder.Add(cmd);
-        }
-        // yieldCount := yieldCount + 1;  assume yieldCount == |ys|;
-        var yc = new Bpl.IdentifierExpr(s.Tok, yieldCountVariable);
-        var incYieldCount = Bpl.Cmd.SimpleAssign(s.Tok, yc, Bpl.Expr.Binary(s.Tok, Bpl.BinaryOperator.Opcode.Add, yc, Bpl.Expr.Literal(1)));
-        builder.Add(incYieldCount);
-        builder.Add(TrAssumeCmd(s.Tok, YieldCountAssumption(iter, etran)));
-        // assume $IsGoodHeap($Heap);
-        builder.Add(AssumeGoodHeap(s.Tok, etran));
-        // assert YieldEnsures[subst];  // where 'subst' replaces "old(E)" with "E" being evaluated in $_OldIterHeap
-        var yeEtran = new ExpressionTranslator(this, predef, etran.HeapExpr, new Bpl.IdentifierExpr(s.Tok, "$_OldIterHeap", predef.HeapType));
-        foreach (var p in iter.YieldEnsures) {
-          bool splitHappened;  // actually, we don't care
-          var ss = TrSplitExpr(p.E, yeEtran, true, out splitHappened);
-          foreach (var split in ss) {
-            if (RefinementToken.IsInherited(split.E.tok, currentModule)) {
-              // this postcondition was inherited into this module, so just ignore it
-            } else if (split.IsChecked) {
-              var yieldToken = new NestedToken(s.Tok, split.E.tok);
-              builder.Add(AssertNS(yieldToken, split.E, "possible violation of yield-ensures condition", stmt.Tok, null));
-            }
-          }
-          builder.Add(TrAssumeCmd(stmt.Tok, yeEtran.TrExpr(p.E)));
-        }
-        YieldHavoc(iter.tok, iter, builder, etran);
-        builder.Add(CaptureState(s));
-
-      } else if (stmt is AssignSuchThatStmt) {
-        var s = (AssignSuchThatStmt)stmt;
-        AddComment(builder, s, "assign-such-that statement");
-        // Essentially, treat like an assert (that values for the LHS exist), a havoc (of the LHS), and an
-        // assume (of the RHS).  However, we also need to check the well-formedness of the LHS and RHS.
-        // The well-formedness of the LHS is done by the havoc. The well-formedness of the RHS is most
-        // easily done after that havoc, but we need to be careful about two things:
-        //   - We should not generate any errors for uses of LHSs. This is not an issue for fields or
-        //     array elements, because they already have values before reaching the assign-such-that statement.
-        //     (Note, "this.f" is not allowed as a LHS in the first division of a constructor.)
-        //     For any local variable or out-parameter x that's a LHS, we create a new variable x' and
-        //     substitute x' for x in the RHS before doing the well-formedness check.
-        //   - The well-formedness checks need to be able to assume that each x' has a value of its
-        //     type. However, this assumption must not carry over to the existence assertion, because
-        //     then everything will be provable if x' is of a known-empty type. Instead, the well-formedness
-        //     check is wrapped inside an "if" whose guard is the type antecedent. After the existence
-        //     check, the type antecedent is assumed of the original x, the RHS is assumed, and x is
-        //     marked off has having been definitely assigned.
-        //
-        // So, the Dafny statement
-        //     E.f, x :| RHS(x);
-        // is translated into the following Boogie code:
-        //     var x';
-        //     Tr[[ E.f := *; ]]  // this havoc is translated like a Dafny assignment statement, which means
-        //                        // the well-formedness of E is checked here
-        //     if (typeAntecedent(x')) {
-        //       check well-formedness of RHS(x');
-        //     }
-        //     assert (exists x'' :: RHS(x''));  // for ":| assume", omit this line; for ":|", LHS is only allowed to contain simple variables
-        //     defass$x := true;
-        //     havoc x;
-        //     assume RHS(x);
-
-        var simpleLHSs = new List<IdentifierExpr>();
-        Bpl.Expr typeAntecedent = Bpl.Expr.True;
-        var havocLHSs = new List<Expression>();
-        var havocRHSs = new List<AssignmentRhs>();
-        var substMap = new Dictionary<IVariable, Expression>();
-        foreach (var lhs in s.Lhss) {
-          var lvalue = lhs.Resolved;
-          if (lvalue is IdentifierExpr ide) {
-            simpleLHSs.Add(ide);
-            var wh = SetupVariableAsLocal(ide.Var, substMap, builder, locals, etran);
-            typeAntecedent = BplAnd(typeAntecedent, wh);
-          } else {
-            havocLHSs.Add(lhs.Resolved);
-            havocRHSs.Add(new HavocRhs(lhs.tok));  // note, a HavocRhs is constructed as already resolved
-          }
-        }
-        ProcessLhss(havocLHSs, false, true, builder, locals, etran, out var lhsBuilder, out var bLhss, out _, out _, out _);
-        ProcessRhss(lhsBuilder, bLhss, havocLHSs, havocRHSs, builder, locals, etran);
-
-        // Here comes the well-formedness check
-        var wellFormednessBuilder = new BoogieStmtListBuilder(this);
-        var rhs = Substitute(s.Expr, null, substMap);
-        TrStmt_CheckWellformed(rhs, wellFormednessBuilder, locals, etran, false);
-        var ifCmd = new Bpl.IfCmd(s.Tok, typeAntecedent, wellFormednessBuilder.Collect(s.Tok), null, null);
-        builder.Add(ifCmd);
-
-        // Here comes the assert part
-        if (s.AssumeToken == null) {
-          substMap = new Dictionary<IVariable, Expression>();
-          var bvars = new List<BoundVar>();
-          foreach (var lhs in s.Lhss) {
-            var l = lhs.Resolved;
-            if (l is IdentifierExpr x) {
-              CloneVariableAsBoundVar(x.tok, x.Var, "$as#" + x.Name, out var bv, out var ie);
-              bvars.Add(bv);
-              substMap.Add(x.Var, ie);
-            } else {
-              // other forms of LHSs have been ruled out by the resolver (it would be possible to
-              // handle them, but it would involve heap-update expressions--the translation would take
-              // effort, and it's not certain that the existential would be successful in verification).
-              Contract.Assume(false);  // unexpected case
-            }
-          }
-
-          GenerateAndCheckGuesses(s.Tok, bvars, s.Bounds, Substitute(s.Expr, null, substMap), TrTrigger(etran, s.Attributes, s.Tok, substMap), builder, etran);
-        }
-
-        // Mark off the simple variables as having definitely been assigned AND THEN havoc their values. By doing them
-        // in this order, they type antecedents will in effect be assumed.
-        var bHavocLHSs = new List<Bpl.IdentifierExpr>();
-        foreach (var lhs in simpleLHSs) {
-          MarkDefiniteAssignmentTracker(lhs, builder);
-          bHavocLHSs.Add((Bpl.IdentifierExpr)etran.TrExpr(lhs));
-        }
-        builder.Add(new Bpl.HavocCmd(s.Tok, bHavocLHSs));
-
-        // End by doing the assume
-        builder.Add(TrAssumeCmd(s.Tok, etran.TrExpr(s.Expr)));
-        builder.Add(CaptureState(s));  // just do one capture state--here, at the very end (that is, don't do one before the assume)
-
-      } else if (stmt is UpdateStmt) {
-        var s = (UpdateStmt)stmt;
-        // This UpdateStmt can be single-target assignment, a multi-assignment, a call statement, or
-        // an array-range update.  Handle the multi-assignment here and handle the others as for .ResolvedStatements.
-        var resolved = s.ResolvedStatements;
-        if (resolved.Count == 1) {
-          TrStmt(resolved[0], builder, locals, etran);
-        } else {
-          AddComment(builder, s, "update statement");
-          var lhss = new List<Expression>();
-          foreach (var lhs in s.Lhss) {
-            lhss.Add(lhs.Resolved);
-          }
-          List<AssignToLhs> lhsBuilder;
-          List<Bpl.IdentifierExpr> bLhss;
-          // note: because we have more than one expression, we always must assign to Boogie locals in a two
-          // phase operation. Thus rhssCanAffectPreviouslyKnownExpressions is just true.
-          Contract.Assert(1 < lhss.Count);
-
-          Bpl.Expr[] lhsObjs, lhsFields;
-          string[] lhsNames;
-          ProcessLhss(lhss, true, false, builder, locals, etran, out lhsBuilder, out bLhss, out lhsObjs, out lhsFields, out lhsNames);
-          // We know that, because the translation saves to a local variable, that the RHS always need to
-          // generate a new local, i.e. bLhss is just all nulls.
-          Contract.Assert(Contract.ForAll(bLhss, lhs => lhs == null));
-          // This generates the assignments, and gives them to us as finalRhss.
-          var finalRhss = ProcessUpdateAssignRhss(lhss, s.Rhss, builder, locals, etran);
-          // ProcessLhss has laid down framing conditions and the ProcessUpdateAssignRhss will check subranges (nats),
-          // but we need to generate the distinctness condition (two LHS are equal only when the RHS is also
-          // equal). We need both the LHS and the RHS to do this, which is why we need to do it here.
-          CheckLhssDistinctness(finalRhss, s.Rhss, lhss, builder, etran, lhsObjs, lhsFields, lhsNames, s.OriginalInitialLhs);
-          // Now actually perform the assignments to the LHS.
-          for (int i = 0; i < lhss.Count; i++) {
-            lhsBuilder[i](finalRhss[i], s.Rhss[i] is HavocRhs, builder, etran);
-          }
-          builder.Add(CaptureState(s));
-        }
-
-      } else if (stmt is AssignOrReturnStmt) {
-        AddComment(builder, stmt, "assign-or-return statement (:-)");
-        AssignOrReturnStmt s = (AssignOrReturnStmt)stmt;
-        TrStmtList(s.ResolvedStatements, builder, locals, etran);
-
-      } else if (stmt is AssignStmt) {
-        AddComment(builder, stmt, "assignment statement");
-        AssignStmt s = (AssignStmt)stmt;
-        TrAssignment(stmt, s.Lhs.Resolved, s.Rhs, builder, locals, etran);
-
-      } else if (stmt is CallStmt) {
-        AddComment(builder, stmt, "call statement");
-        TrCallStmt((CallStmt)stmt, builder, locals, etran, null);
-
-      } else if (stmt is DividedBlockStmt) {
-        var s = (DividedBlockStmt)stmt;
-        AddComment(builder, stmt, "divided block before new;");
-        var prevDefiniteAssignmentTrackerCount = definiteAssignmentTrackers.Count;
-        var tok = s.SeparatorTok ?? s.Tok;
-        // a DividedBlockStmt occurs only inside a Constructor body of a class
-        var cl = (ClassDecl)((Constructor)codeContext).EnclosingClass;
-        var fields = Concat(cl.InheritedMembers, cl.Members).ConvertAll(member =>
-          member is Field && !member.IsStatic && !member.IsInstanceIndependentConstant ? (Field)member : null);
-        fields.RemoveAll(f => f == null);
-        var localSurrogates = fields.ConvertAll(f => new Bpl.LocalVariable(f.tok, new TypedIdent(f.tok, SurrogateName(f), TrType(f.Type))));
-        locals.AddRange(localSurrogates);
-        fields.Iter(f => AddDefiniteAssignmentTrackerSurrogate(f, cl, locals));
-
-        Contract.Assert(!inBodyInitContext);
-        inBodyInitContext = true;
-        TrStmtList(s.BodyInit, builder, locals, etran);
-        Contract.Assert(inBodyInitContext);
-        inBodyInitContext = false;
-
-        // The "new;" translates into an allocation of "this"
-        AddComment(builder, stmt, "new;");
-        fields.Iter(f => CheckDefiniteAssignmentSurrogate(s.SeparatorTok ?? s.EndTok, f, true, builder));
-        fields.Iter(f => RemoveDefiniteAssignmentTrackerSurrogate(f));
-        var th = new ThisExpr(cl);
-        var bplThis = (Bpl.IdentifierExpr)etran.TrExpr(th);
-        SelectAllocateObject(tok, bplThis, th.Type, false, builder, etran);
-        for (int i = 0; i < fields.Count; i++) {
-          // assume $Heap[this, f] == this.f;
-          var mse = new MemberSelectExpr(tok, th, fields[i]);
-          Bpl.Expr surr = new Bpl.IdentifierExpr(tok, localSurrogates[i]);
-          surr = CondApplyUnbox(tok, surr, fields[i].Type, mse.Type);
-          builder.Add(new Bpl.AssumeCmd(tok, Bpl.Expr.Eq(etran.TrExpr(mse), surr)));
-        }
-        CommitAllocatedObject(tok, bplThis, null, builder, etran);
-
-        AddComment(builder, stmt, "divided block after new;");
-        TrStmtList(s.BodyProper, builder, locals, etran);
-        RemoveDefiniteAssignmentTrackers(s.Body, prevDefiniteAssignmentTrackerCount);
-
-      } else if (stmt is BlockStmt) {
-        var s = (BlockStmt)stmt;
-        var prevDefiniteAssignmentTrackerCount = definiteAssignmentTrackers.Count;
-        TrStmtList(s.Body, builder, locals, etran);
-        RemoveDefiniteAssignmentTrackers(s.Body, prevDefiniteAssignmentTrackerCount);
-      } else if (stmt is IfStmt) {
-        AddComment(builder, stmt, "if statement");
-        IfStmt s = (IfStmt)stmt;
-        Expression guard;
-        if (s.Guard == null) {
-          guard = null;
-        } else {
-          guard = s.IsBindingGuard ? AlphaRename((ExistsExpr)s.Guard, "eg$") : s.Guard;
-          TrStmt_CheckWellformed(guard, builder, locals, etran, true);
-        }
-        BoogieStmtListBuilder b = new BoogieStmtListBuilder(this);
-        if (s.IsBindingGuard) {
-          CurrentIdGenerator.Push();
-          var exists = (ExistsExpr)s.Guard;  // the original (that is, not alpha-renamed) guard
-          IntroduceAndAssignExistentialVars(exists, b, builder, locals, etran, stmt.IsGhost);
-          CurrentIdGenerator.Pop();
-        }
-        CurrentIdGenerator.Push();
-        Bpl.StmtList thn = TrStmt2StmtList(b, s.Thn, locals, etran);
-        CurrentIdGenerator.Pop();
-        Bpl.StmtList els;
-        Bpl.IfCmd elsIf = null;
-        b = new BoogieStmtListBuilder(this);
-        if (s.IsBindingGuard) {
-          b.Add(TrAssumeCmd(guard.tok, Bpl.Expr.Not(etran.TrExpr(guard))));
-        }
-        if (s.Els == null) {
-          els = b.Collect(s.Tok);
-        } else {
-          CurrentIdGenerator.Push();
-          els = TrStmt2StmtList(b, s.Els, locals, etran);
-          CurrentIdGenerator.Pop();
-          if (els.BigBlocks.Count == 1) {
-            Bpl.BigBlock bb = els.BigBlocks[0];
-            if (bb.LabelName == null && bb.simpleCmds.Count == 0 && bb.ec is Bpl.IfCmd) {
-              elsIf = (Bpl.IfCmd)bb.ec;
-              els = null;
-            }
-          }
-        }
-        builder.Add(new Bpl.IfCmd(stmt.Tok, guard == null || s.IsBindingGuard ? null : etran.TrExpr(guard), thn, elsIf, els));
-
-      } else if (stmt is AlternativeStmt) {
-        AddComment(builder, stmt, "alternative statement");
-        var s = (AlternativeStmt)stmt;
-        var elseCase = Assert(s.Tok, Bpl.Expr.False, "alternative cases fail to cover all possibilties");
-        TrAlternatives(s.Alternatives, elseCase, null, builder, locals, etran, stmt.IsGhost);
-
-      } else if (stmt is WhileStmt) {
-        AddComment(builder, stmt, "while statement");
-        this.fuelContext = FuelSetting.ExpandFuelContext(stmt.Attributes, stmt.Tok, this.fuelContext, this.reporter);
-        DefineFuelConstant(stmt.Tok, stmt.Attributes, builder, etran);
-        var s = (WhileStmt)stmt;
-        BodyTranslator bodyTr = null;
-        if (s.Body != null) {
-          bodyTr = delegate(BoogieStmtListBuilder bld, ExpressionTranslator e) {
-            CurrentIdGenerator.Push();
-            TrStmt(s.Body, bld, locals, e);
-            CurrentIdGenerator.Pop();
-          };
-        }
-        TrLoop(s, s.Guard, bodyTr, builder, locals, etran);
-        this.fuelContext = FuelSetting.PopFuelContext();
-      } else if (stmt is AlternativeLoopStmt) {
-        AddComment(builder, stmt, "alternative loop statement");
-        var s = (AlternativeLoopStmt)stmt;
-        var tru = new LiteralExpr(s.Tok, true);
-        tru.Type = Type.Bool; // resolve here
-        TrLoop(s, tru,
-          delegate(BoogieStmtListBuilder bld, ExpressionTranslator e) {
-            TrAlternatives(s.Alternatives, null, new Bpl.BreakCmd(s.Tok, null), bld, locals, e, stmt.IsGhost);
-          },
-          builder, locals, etran);
-
-      } else if (stmt is ForLoopStmt) {
-        var s = (ForLoopStmt)stmt;
-        AddComment(builder, stmt, "for-loop statement");
-
-        var indexVar = s.LoopIndex;
-        var indexVarName = indexVar.AssignUniqueName(currentDeclaration.IdGenerator);
-        var dIndex = new IdentifierExpr(indexVar.tok, indexVar);
-        var bIndexVar = new Bpl.LocalVariable(indexVar.tok, new Bpl.TypedIdent(indexVar.Tok, indexVarName, TrType(indexVar.Type)));
-        locals.Add(bIndexVar);
-        var bIndex = new Bpl.IdentifierExpr(indexVar.tok, indexVarName);
-
-        var lo = s.GoingUp ? s.Start : s.End;
-        var hi = s.GoingUp ? s.End : s.Start;
-        Expression dLo = null;
-        Expression dHi = null;
-        Bpl.IdentifierExpr bLo = null;
-        Bpl.IdentifierExpr bHi = null;
-        if (lo != null) {
-          var name = indexVarName + "#lo";
-          var bLoVar = new Bpl.LocalVariable(lo.tok, new Bpl.TypedIdent(lo.tok, name, Bpl.Type.Int));
-          locals.Add(bLoVar);
-          bLo = new Bpl.IdentifierExpr(lo.tok, name);
-          CheckWellformed(lo, new WFOptions(null, false), locals, builder, etran);
-          builder.Add(Bpl.Cmd.SimpleAssign(lo.tok, bLo, etran.TrExpr(lo)));
-          dLo = new BoogieWrapper(bLo, lo.Type);
-        }
-        if (hi != null) {
-          var name = indexVarName + "#hi";
-          var bHiVar = new Bpl.LocalVariable(hi.tok, new Bpl.TypedIdent(hi.tok, name, Bpl.Type.Int));
-          locals.Add(bHiVar);
-          bHi = new Bpl.IdentifierExpr(hi.tok, name);
-          CheckWellformed(hi, new WFOptions(null, false), locals, builder, etran);
-          builder.Add(Bpl.Cmd.SimpleAssign(hi.tok, bHi, etran.TrExpr(hi)));
-          dHi = new BoogieWrapper(bHi, hi.Type);
-        }
-
-        // check lo <= hi
-        if (lo != null && hi != null) {
-          builder.Add(Assert(lo.tok, Bpl.Expr.Le(bLo, bHi), "lower bound must not exceed upper bound"));
-        }
-        // check forall x :: lo <= x <= hi ==> Is(x, typ)
-        {
-          // The check, if needed, is performed like this:
-          //   var x: int;
-          //   havoc x;
-          //   assume lo <= x <= hi;
-          //   assert Is(x, typ);
-          var tok = indexVar.tok;
-          var name = indexVarName + "#x";
-          var xVar = new Bpl.LocalVariable(tok, new Bpl.TypedIdent(tok, name, Bpl.Type.Int));
-          var x = new Bpl.IdentifierExpr(tok, name);
-          string msg;
-          var cre = GetSubrangeCheck(x, Type.Int, indexVar.Type, out msg);
-          if (cre != null) {
-            locals.Add(xVar);
-            builder.Add(new Bpl.HavocCmd(tok, new List<Bpl.IdentifierExpr>() { x }));
-            builder.Add(new Bpl.AssumeCmd(tok, ForLoopBounds(x, bLo, bHi)));
-            builder.Add(Assert(tok, cre, "entire range must be assignable to index variable, but some " + msg));
-          }
-        }
-
-        // initialize the index variable
-        builder.Add(Bpl.Cmd.SimpleAssign(indexVar.tok, bIndex, s.GoingUp ? bLo : bHi));
-
-        // build the guard expression
-        Expression guard;
-        if (lo == null || hi == null) {
-          guard = LiteralExpr.CreateBoolLiteral(s.Tok, true);
-        } else {
-          guard = Expression.CreateNot(s.Tok, Expression.CreateEq(dIndex, s.GoingUp ? dHi : dLo, indexVar.Type));
-        }
-
-        // free invariant lo <= i <= hi
-        var freeInvariant = ForLoopBounds(bIndex, bLo, bHi);
-
-        BodyTranslator bodyTr = null;
-        if (s.Body != null) {
-          bodyTr = delegate(BoogieStmtListBuilder bld, ExpressionTranslator e) {
-            CurrentIdGenerator.Push();
-            if (!s.GoingUp) {
-              bld.Add(Bpl.Cmd.SimpleAssign(s.Tok, bIndex, Bpl.Expr.Sub(bIndex, Bpl.Expr.Literal(1))));
-            }
-            TrStmt(s.Body, bld, locals, e);
-            if (s.GoingUp) {
-              bld.Add(Bpl.Cmd.SimpleAssign(s.Tok, bIndex, Bpl.Expr.Add(bIndex, Bpl.Expr.Literal(1))));
-            }
-            CurrentIdGenerator.Pop();
-          };
-        }
-
-        TrLoop(s, guard, bodyTr, builder, locals, etran, freeInvariant, s.Decreases.Expressions.Count != 0);
-
-      } else if (stmt is ModifyStmt) {
-        AddComment(builder, stmt, "modify statement");
-        var s = (ModifyStmt)stmt;
-        // check well-formedness of the modifies clauses
-        CheckFrameWellFormed(new WFOptions(), s.Mod.Expressions, locals, builder, etran);
-        // check that the modifies is a subset
-        CheckFrameSubset(s.Tok, s.Mod.Expressions, null, null, etran, builder, "modify statement may violate context's modifies clause", null);
-        // cause the change of the heap according to the given frame
-        var suffix = CurrentIdGenerator.FreshId("modify#");
-        string modifyFrameName = "$Frame$" + suffix;
-        var preModifyHeapVar = new Bpl.LocalVariable(s.Tok, new Bpl.TypedIdent(s.Tok, "$PreModifyHeap$" + suffix, predef.HeapType));
-        locals.Add(preModifyHeapVar);
-        DefineFrame(s.Tok, s.Mod.Expressions, builder, locals, modifyFrameName);
-        if (s.Body == null) {
-          var preModifyHeap = new Bpl.IdentifierExpr(s.Tok, preModifyHeapVar);
-          // preModifyHeap := $Heap;
-          builder.Add(Bpl.Cmd.SimpleAssign(s.Tok, preModifyHeap, etran.HeapExpr));
-          // havoc $Heap;
-          builder.Add(new Bpl.HavocCmd(s.Tok, new List<Bpl.IdentifierExpr> { (Bpl.IdentifierExpr/*TODO: this cast is rather dubious*/)etran.HeapExpr }));
-          // assume $HeapSucc(preModifyHeap, $Heap);   OR $HeapSuccGhost
-          builder.Add(TrAssumeCmd(s.Tok, HeapSucc(preModifyHeap, etran.HeapExpr, s.IsGhost)));
-          // assume nothing outside the frame was changed
-          var etranPreLoop = new ExpressionTranslator(this, predef, preModifyHeap);
-          var updatedFrameEtran = new ExpressionTranslator(etran, modifyFrameName);
-          builder.Add(TrAssumeCmd(s.Tok, FrameConditionUsingDefinedFrame(s.Tok, etranPreLoop, etran, updatedFrameEtran)));
-        } else {
-          // do the body, but with preModifyHeapVar as the governing frame
-          var updatedFrameEtran = new ExpressionTranslator(etran, modifyFrameName);
-          CurrentIdGenerator.Push();
-          TrStmt(s.Body, builder, locals, updatedFrameEtran);
-          CurrentIdGenerator.Pop();
-        }
-        builder.Add(CaptureState(stmt));
-
-      } else if (stmt is ForallStmt) {
-        var s = (ForallStmt)stmt;
-        this.fuelContext = FuelSetting.ExpandFuelContext(stmt.Attributes, stmt.Tok, this.fuelContext, this.reporter);
-
-        CurrentIdGenerator.Push();
-        if (s.Kind == ForallStmt.BodyKind.Assign) {
-          AddComment(builder, stmt, "forall statement (assign)");
-          Contract.Assert(s.Ens.Count == 0);
-          if (s.BoundVars.Count == 0) {
-            TrStmt(s.Body, builder, locals, etran);
-          } else {
-            var s0 = (AssignStmt)s.S0;
-            var definedness = new BoogieStmtListBuilder(this);
-            var updater = new BoogieStmtListBuilder(this);
-            DefineFuelConstant(stmt.Tok, stmt.Attributes, definedness, etran);
-            TrForallAssign(s, s0, definedness, updater, locals, etran);
-            // All done, so put the two pieces together
-            builder.Add(new Bpl.IfCmd(s.Tok, null, definedness.Collect(s.Tok), null, updater.Collect(s.Tok)));
-            builder.Add(CaptureState(stmt));
-          }
-
-        } else if (s.Kind == ForallStmt.BodyKind.Call) {
-          AddComment(builder, stmt, "forall statement (call)");
-          Contract.Assert(s.Ens.Count == 0);
-          if (s.BoundVars.Count == 0) {
-            Contract.Assert(LiteralExpr.IsTrue(s.Range));  // follows from the invariant of ForallStmt
-            TrStmt(s.Body, builder, locals, etran);
-          } else {
-            var s0 = (CallStmt)s.S0;
-            if (Attributes.Contains(s.Attributes, "_trustWellformed")) {
-              TrForallStmtCall(s.Tok, s.BoundVars, s.Bounds, s.Range, null, s.ForallExpressions, s0, null, builder, locals, etran);
-            } else {
-              var definedness = new BoogieStmtListBuilder(this);
-              DefineFuelConstant(stmt.Tok, stmt.Attributes, definedness, etran);
-              var exporter = new BoogieStmtListBuilder(this);
-              TrForallStmtCall(s.Tok, s.BoundVars, s.Bounds, s.Range, null, s.ForallExpressions, s0, definedness, exporter, locals, etran);
-              // All done, so put the two pieces together
-              builder.Add(new Bpl.IfCmd(s.Tok, null, definedness.Collect(s.Tok), null, exporter.Collect(s.Tok)));
-            }
-            builder.Add(CaptureState(stmt));
-          }
-
-        } else if (s.Kind == ForallStmt.BodyKind.Proof) {
-          AddComment(builder, stmt, "forall statement (proof)");
-          var definedness = new BoogieStmtListBuilder(this);
-          var exporter = new BoogieStmtListBuilder(this);
-          DefineFuelConstant(stmt.Tok, stmt.Attributes, definedness, etran);
-          TrForallProof(s, definedness, exporter, locals, etran);
-          // All done, so put the two pieces together
-          builder.Add(new Bpl.IfCmd(s.Tok, null, definedness.Collect(s.Tok), null, exporter.Collect(s.Tok)));
-          builder.Add(CaptureState(stmt));
-
-        } else {
-          Contract.Assert(false);  // unexpected kind
-        }
-        CurrentIdGenerator.Pop();
-        this.fuelContext = FuelSetting.PopFuelContext();
-      } else if (stmt is CalcStmt) {
-        /* Translate into:
-        if (*) {
-            assert wf(line0);
-        } else if (*) {
-            assume wf(line0);
-            // if op is ==>: assume line0;
-            hint0;
-            assert wf(line1);
-            assert line0 op line1;
-            assume false;
-        } else if (*) { ...
-        } else if (*) {
-            assume wf(line<n-1>);
-            // if op is ==>: assume line<n-1>;
-            hint<n-1>;
-            assert wf(line<n>);
-            assert line<n-1> op line<n>;
-            assume false;
-        }
-        assume line<0> op line<n>;
-        */
-        var s = (CalcStmt)stmt;
-        Contract.Assert(s.Steps.Count == s.Hints.Count); // established by the resolver
-        AddComment(builder, stmt, "calc statement");
-        this.fuelContext = FuelSetting.ExpandFuelContext(stmt.Attributes, stmt.Tok, this.fuelContext, this.reporter);
-        DefineFuelConstant(stmt.Tok, stmt.Attributes, builder, etran);
-        CurrentIdGenerator.Push();  // put the entire calc statement within its own sub-branch
-        if (s.Lines.Count > 0) {
-          Bpl.IfCmd ifCmd = null;
-          BoogieStmtListBuilder b;
-          // if the dangling hint is empty, do not generate anything for the dummy step
-          var stepCount = s.Hints.Last().Body.Count == 0 ? s.Steps.Count - 1 : s.Steps.Count;
-          // check steps:
-          for (int i = stepCount; 0 <= --i; ) {
-            b = new BoogieStmtListBuilder(this);
-            // assume wf[line<i>]:
-            AddComment(b, stmt, "assume wf[lhs]");
-            CurrentIdGenerator.Push();
-            assertAsAssume = true;
-            TrStmt_CheckWellformed(CalcStmt.Lhs(s.Steps[i]), b, locals, etran, false);
-            assertAsAssume = false;
-            if (s.Steps[i] is BinaryExpr && (((BinaryExpr)s.Steps[i]).ResolvedOp == BinaryExpr.ResolvedOpcode.Imp)) {
-              // assume line<i>:
-              AddComment(b, stmt, "assume lhs");
-              b.Add(TrAssumeCmd(s.Tok, etran.TrExpr(CalcStmt.Lhs(s.Steps[i]))));
-            }
-            // hint:
-            AddComment(b, stmt, "Hint" + i.ToString());
-            TrStmt(s.Hints[i], b, locals, etran);
-            if (i < s.Steps.Count - 1) { // non-dummy step
-              // check well formedness of the goal line:
-              AddComment(b, stmt, "assert wf[rhs]");
-              if (s.Steps[i] is TernaryExpr) {
-                // check the prefix-equality limit
-                var index = ((TernaryExpr) s.Steps[i]).E0;
-                TrStmt_CheckWellformed(index, b, locals, etran, false);
-                if (index.Type.IsNumericBased(Type.NumericPersuasion.Int)) {
-                  b.Add(AssertNS(index.tok, Bpl.Expr.Le(Bpl.Expr.Literal(0), etran.TrExpr(index)), "prefix-equality limit must be at least 0"));
-                }
-              }
-              TrStmt_CheckWellformed(CalcStmt.Rhs(s.Steps[i]), b, locals, etran, false);
-              bool splitHappened;
-              var ss = TrSplitExpr(s.Steps[i], etran, true, out splitHappened);
-              // assert step:
-              AddComment(b, stmt, "assert line" + i.ToString() + " " + (s.StepOps[i] ?? s.Op).ToString() + " line" + (i + 1).ToString());
-              if (!splitHappened) {
-                b.Add(AssertNS(s.Lines[i + 1].tok, etran.TrExpr(s.Steps[i]), "the calculation step between the previous line and this line might not hold"));
-              } else {
-                foreach (var split in ss) {
-                  if (split.IsChecked) {
-                    b.Add(AssertNS(s.Lines[i + 1].tok, split.E, "the calculation step between the previous line and this line might not hold"));
-                  }
-                }
-              }
-            }
-            b.Add(TrAssumeCmd(s.Tok, Bpl.Expr.False));
-            ifCmd = new Bpl.IfCmd(s.Tok, null, b.Collect(s.Tok), ifCmd, null);
-            CurrentIdGenerator.Pop();
-          }
-          // check well formedness of the first line:
-          b = new BoogieStmtListBuilder(this);
-          AddComment(b, stmt, "assert wf[initial]");
-          Contract.Assert(s.Result != null); // established by the resolver
-          TrStmt_CheckWellformed(CalcStmt.Lhs(s.Result), b, locals, etran, false);
-          b.Add(TrAssumeCmd(s.Tok, Bpl.Expr.False));
-          ifCmd = new Bpl.IfCmd(s.Tok, null, b.Collect(s.Tok), ifCmd, null);
-          builder.Add(ifCmd);
-          // assume result:
-          if (s.Steps.Count > 1) {
-            builder.Add(TrAssumeCmd(s.Tok, etran.TrExpr(s.Result)));
-          }
-        }
-        CurrentIdGenerator.Pop();
-        this.fuelContext = FuelSetting.PopFuelContext();
-      } else if (stmt is ConcreteSyntaxStatement) {
-        ConcreteSyntaxStatement s = (ConcreteSyntaxStatement)stmt;
-        TrStmt(s.ResolvedStatement, builder, locals, etran);
-      } else if (stmt is MatchStmt) {
-        var s = (MatchStmt)stmt;
-        TrStmt_CheckWellformed(s.Source, builder, locals, etran, true);
-        Bpl.Expr source = etran.TrExpr(s.Source);
-        var b = new BoogieStmtListBuilder(this);
-        b.Add(TrAssumeCmd(stmt.Tok, Bpl.Expr.False));
-        Bpl.StmtList els = b.Collect(stmt.Tok);
-        Bpl.IfCmd ifCmd = null;
-        foreach (var missingCtor in s.MissingCases) {
-          // havoc all bound variables
-          b = new BoogieStmtListBuilder(this);
-          List<Variable> newLocals = new List<Variable>();
-          Bpl.Expr r = CtorInvocation(s.Tok, missingCtor, etran, newLocals, b);
-          locals.AddRange(newLocals);
-
-          if (newLocals.Count != 0)
-          {
-            List<Bpl.IdentifierExpr> havocIds = new List<Bpl.IdentifierExpr>();
-            foreach (Variable local in newLocals) {
-              havocIds.Add(new Bpl.IdentifierExpr(local.tok, local));
-            }
-            builder.Add(new Bpl.HavocCmd(s.Tok, havocIds));
-          }
-          String missingStr = s.Context.FillHole(new IdCtx(new KeyValuePair<string, DatatypeCtor>(missingCtor.Name, missingCtor))).AbstractAllHoles().ToString();
-          b.Add(Assert(s.Tok, Bpl.Expr.False, "missing case in match statement: " + missingStr));
-
-          Bpl.Expr guard = Bpl.Expr.Eq(source, r);
-          ifCmd = new Bpl.IfCmd(s.Tok, guard, b.Collect(s.Tok), ifCmd, els);
-          els = null;
-        }
-        for (int i = s.Cases.Count; 0 <= --i; ) {
-          var mc = (MatchCaseStmt)s.Cases[i];
-          CurrentIdGenerator.Push();
-          // havoc all bound variables
-          b = new BoogieStmtListBuilder(this);
-          List<Variable> newLocals = new List<Variable>();
-          Bpl.Expr r = CtorInvocation(mc, s.Source.Type, etran, newLocals, b, s.IsGhost ? NOALLOC : ISALLOC);
-          locals.AddRange(newLocals);
-
-          if (newLocals.Count != 0)
-          {
-            List<Bpl.IdentifierExpr> havocIds = new List<Bpl.IdentifierExpr>();
-            foreach (Variable local in newLocals) {
-              havocIds.Add(new Bpl.IdentifierExpr(local.tok, local));
-            }
-            builder.Add(new Bpl.HavocCmd(mc.tok, havocIds));
-          }
-
-          // translate the body into b
-          var prevDefiniteAssignmentTrackerCount = definiteAssignmentTrackers.Count;
-          TrStmtList(mc.Body, b, locals, etran);
-          RemoveDefiniteAssignmentTrackers(mc.Body, prevDefiniteAssignmentTrackerCount);
-
-          Bpl.Expr guard = Bpl.Expr.Eq(source, r);
-          ifCmd = new Bpl.IfCmd(mc.tok, guard, b.Collect(mc.tok), ifCmd, els);
-          els = null;
-          CurrentIdGenerator.Pop();
-        }
-        Contract.Assert(ifCmd != null);  // follows from the fact that s.Cases.Count + s.MissingCases.Count != 0.
-        builder.Add(ifCmd);
-
-      } else if (stmt is VarDeclStmt) {
-        var s = (VarDeclStmt)stmt;
-        var newLocalIds = new List<Bpl.IdentifierExpr>();
-        int i = 0;
-        foreach (var local in s.Locals) {
-          Bpl.Type varType = TrType(local.Type);
-          Bpl.Expr wh = GetWhereClause(local.Tok,
-            new Bpl.IdentifierExpr(local.Tok, local.AssignUniqueName(currentDeclaration.IdGenerator), varType),
-            local.Type, etran, isAllocContext.Var(stmt.IsGhost, local));
-          // if needed, register definite-assignment tracking for this local
-          var needDefiniteAssignmentTracking = s.Update == null || s.Update is AssignSuchThatStmt;
-          if (s.Update is UpdateStmt) {
-            // there is an initial assignment, but we need to look out for "*" being that assignment
-            var us = (UpdateStmt)s.Update;
-            if (i < us.Rhss.Count && us.Rhss[i] is HavocRhs) {
-              needDefiniteAssignmentTracking = true;
-            }
-          }
-          if (needDefiniteAssignmentTracking) {
-            var defassExpr = AddDefiniteAssignmentTracker(local, locals);
-            if (wh != null && defassExpr != null) {
-              // make the "where" expression be "defass ==> wh", because we don't want to assume anything
-              // before the variable has been assigned (for a variable that needs definite-assignment tracking
-              // in the first place)
-              wh = BplImp(defassExpr, wh);
-            }
-          }
-          // create the variable itself (now that "wh" may mention the definite-assignment tracker)
-          Bpl.LocalVariable var = new Bpl.LocalVariable(local.Tok, new Bpl.TypedIdent(local.Tok, local.AssignUniqueName(currentDeclaration.IdGenerator), varType, wh));
-          var.Attributes = etran.TrAttributes(local.Attributes, null);
-          newLocalIds.Add(new Bpl.IdentifierExpr(local.Tok, var));
-          locals.Add(var);
-          i++;
-        }
-        if (s.Update == null) {
-          // it is necessary to do a havoc here in order to give the variable the correct allocation state
-          builder.Add(new HavocCmd(s.Tok, newLocalIds));
-        }
-        // processing of "assumption" variables happens after the variable is possibly havocked above
-        foreach (var local in s.Locals) {
-          if (Attributes.Contains(local.Attributes, "assumption")) {
-            Bpl.Type varType = TrType(local.Type);
-            builder.Add(new AssumeCmd(local.Tok, new Bpl.IdentifierExpr(local.Tok, local.AssignUniqueName(currentDeclaration.IdGenerator), varType), new QKeyValue(local.Tok, "assumption_variable_initialization", new List<object>(), null)));
-          }
-        }
-        if (s.Update != null) {
-          TrStmt(s.Update, builder, locals, etran);
-        }
-      } else if (stmt is VarDeclPattern) {
-        var s = (VarDeclPattern)stmt;
-        foreach (var local in s.LocalVars) {
-          Bpl.LocalVariable bvar = new Bpl.LocalVariable(local.Tok, new Bpl.TypedIdent(local.Tok, local.AssignUniqueName(currentDeclaration.IdGenerator), TrType(local.Type)));
-          locals.Add(bvar);
-          var bIe = new Bpl.IdentifierExpr(bvar.tok, bvar);
-          builder.Add(new Bpl.HavocCmd(local.Tok, new List<Bpl.IdentifierExpr> { bIe }));
-          Bpl.Expr wh = GetWhereClause(local.Tok, bIe, local.Type, etran, isAllocContext.Var(stmt.IsGhost, local));
-          if (wh != null) {
-            builder.Add(TrAssumeCmd(local.Tok, wh));
-          }
-        }
-        var varNameGen = CurrentIdGenerator.NestedFreshIdGenerator("let#");
-        var pat = s.LHS;
-        var rhs = s.RHS;
-        var nm = varNameGen.FreshId(string.Format("#{0}#", 0));
-        var r = new Bpl.LocalVariable(pat.tok, new Bpl.TypedIdent(pat.tok, nm, TrType(rhs.Type)));
-        locals.Add(r);
-        var rIe = new Bpl.IdentifierExpr(rhs.tok, r);
-        CheckWellformedWithResult(rhs, new WFOptions(null, false, false), rIe, pat.Expr.Type, locals, builder, etran);
-        CheckCasePatternShape(pat, rIe, rhs.tok, pat.Expr.Type, builder);
-        builder.Add(TrAssumeCmd(pat.tok, Bpl.Expr.Eq(etran.TrExpr(pat.Expr), rIe)));
-      } else {
-        Contract.Assert(false); throw new cce.UnreachableException();  // unexpected statement
-      }
-    }
-
     /// <summary>
     /// Generates a Boogie expression "lo <= x <= hi", but leaving the lo/hi bound if null.
     /// </summary>
@@ -15178,1981 +14285,6 @@ namespace Microsoft.Dafny {
 
       internal IsAllocType Var(bool ghostStmt, NonglobalVariable var) {
         return Var(allVarsGhost || ghostStmt || var.IsGhost);
-      }
-    }
-
-    internal class ExpressionTranslator
-    {
-      // HeapExpr == null ==> translation of pure (no-heap) expression
-      readonly Bpl.Expr _the_heap_expr;
-      public Bpl.Expr HeapExpr {
-        // The increment of Statistics_HeapUses in the following line is a hack and not entirely a good idea.
-        // Not only does one need to be careful not to mention HeapExpr in contracts (in particular, in ObjectInvariant()
-        // below), but also, the debugger may invoke HeapExpr and that will cause an increment as well.
-        get { Statistics_HeapUses++; return _the_heap_expr; }
-      }
-      public readonly PredefinedDecls predef;
-      public readonly Translator translator;
-      public readonly string This;
-      public readonly string modifiesFrame; // the name of the context's frame variable.
-      readonly Function applyLimited_CurrentFunction;
-      public readonly FuelSetting layerInterCluster;
-      public readonly FuelSetting layerIntraCluster = null;  // a value of null says to do the same as for inter-cluster calls
-      public int Statistics_CustomLayerFunctionCount = 0;
-      public int Statistics_HeapAsQuantifierCount = 0;
-      public int Statistics_HeapUses = 0;
-      public readonly bool stripLits = false;
-      [ContractInvariantMethod]
-      void ObjectInvariant()
-      {
-        // In the following line, it is important to use _the_heap_expr directly, rather than HeapExpr, because
-        // the HeapExpr getter has a side effect on Statistics_HeapUses.
-        Contract.Invariant(_the_heap_expr == null || _the_heap_expr is Bpl.OldExpr || _the_heap_expr is Bpl.IdentifierExpr);
-        Contract.Invariant(predef != null);
-        Contract.Invariant(translator != null);
-        Contract.Invariant(This != null);
-        Contract.Invariant(modifiesFrame != null);
-        Contract.Invariant(layerInterCluster != null);
-        Contract.Invariant(0 <= Statistics_CustomLayerFunctionCount);
-      }
-
-      /// <summary>
-      /// This is the most general constructor.  It is private and takes all the parameters.  Whenever
-      /// one ExpressionTranslator is constructed from another, unchanged parameters are just copied in.
-      /// </summary>
-      ExpressionTranslator(Translator translator, PredefinedDecls predef, Bpl.Expr heap, string thisVar,
-        Function applyLimited_CurrentFunction, FuelSetting layerInterCluster, FuelSetting layerIntraCluster, string modifiesFrame, bool stripLits) {
-
-        Contract.Requires(translator != null);
-        Contract.Requires(predef != null);
-        Contract.Requires(thisVar != null);
-        Contract.Requires(modifiesFrame != null);
-
-        this.translator = translator;
-        this.predef = predef;
-        this._the_heap_expr = heap;
-        this.This = thisVar;
-        this.applyLimited_CurrentFunction = applyLimited_CurrentFunction;
-        this.layerInterCluster = layerInterCluster;
-        if (layerIntraCluster == null) {
-          this.layerIntraCluster = layerInterCluster;
-        } else {
-          this.layerIntraCluster = layerIntraCluster;
-        }
-        this.modifiesFrame = modifiesFrame;
-        this.stripLits = stripLits;
-      }
-
-      public ExpressionTranslator(Translator translator, PredefinedDecls predef, IToken heapToken)
-        : this(translator, predef, new Bpl.IdentifierExpr(heapToken, predef.HeapVarName, predef.HeapType)) {
-        Contract.Requires(translator != null);
-        Contract.Requires(predef != null);
-        Contract.Requires(heapToken != null);
-      }
-
-      public ExpressionTranslator(Translator translator, PredefinedDecls predef, Bpl.Expr heap)
-        : this(translator, predef, heap, "this") {
-        Contract.Requires(translator != null);
-        Contract.Requires(predef != null);
-      }
-
-      public ExpressionTranslator(Translator translator, PredefinedDecls predef, Bpl.Expr heap, Bpl.Expr oldHeap)
-        : this(translator, predef, heap, "this") {
-        Contract.Requires(translator != null);
-        Contract.Requires(predef != null);
-        Contract.Requires(oldHeap != null);
-
-        var old = new ExpressionTranslator(translator, predef, oldHeap);
-        old.oldEtran = old;
-        this.oldEtran = old;
-      }
-
-      public ExpressionTranslator(Translator translator, PredefinedDecls predef, Bpl.Expr heap, string thisVar)
-        : this(translator, predef, heap, thisVar, null, new FuelSetting(translator, 1), null, "$_Frame", false) {
-        Contract.Requires(translator != null);
-        Contract.Requires(predef != null);
-        Contract.Requires(thisVar != null);
-      }
-
-      public ExpressionTranslator(ExpressionTranslator etran, Bpl.Expr heap)
-        : this(etran.translator, etran.predef, heap, etran.This, etran.applyLimited_CurrentFunction, etran.layerInterCluster, etran.layerIntraCluster, etran.modifiesFrame, etran.stripLits)
-      {
-        Contract.Requires(etran != null);
-      }
-
-      public ExpressionTranslator(ExpressionTranslator etran, string modifiesFrame)
-        : this(etran.translator, etran.predef, etran.HeapExpr, etran.This, etran.applyLimited_CurrentFunction, etran.layerInterCluster, etran.layerIntraCluster, modifiesFrame, etran.stripLits) {
-        Contract.Requires(etran != null);
-        Contract.Requires(modifiesFrame != null);
-      }
-
-      ExpressionTranslator oldEtran;
-      public ExpressionTranslator Old {
-        get {
-          Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-          if (oldEtran == null) {
-            oldEtran = new ExpressionTranslator(translator, predef, new Bpl.OldExpr(HeapExpr.tok, HeapExpr), This, applyLimited_CurrentFunction, layerInterCluster, layerIntraCluster, modifiesFrame, stripLits);
-            oldEtran.oldEtran = oldEtran;
-          }
-          return oldEtran;
-        }
-      }
-
-      public ExpressionTranslator OldAt(Label/*?*/ label) {
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-        if (label == null) {
-          return Old;
-        }
-        var heapAt = new Bpl.IdentifierExpr(Token.NoToken, "$Heap_at_" + label.AssignUniqueId(translator.CurrentIdGenerator), predef.HeapType);
-        return new ExpressionTranslator(translator, predef, heapAt, This, applyLimited_CurrentFunction, layerInterCluster, layerIntraCluster, modifiesFrame, stripLits);
-      }
-
-      public bool UsesOldHeap {
-        get {
-          return HeapExpr is Bpl.OldExpr || (HeapExpr is Bpl.IdentifierExpr ide && ide.Name.StartsWith("$Heap_at_"));
-        }
-      }
-
-      public ExpressionTranslator WithLayer(Bpl.Expr layerArgument)
-      {
-        // different layer and 0 fuel amount.
-        Contract.Requires(layerArgument != null);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, null, new FuelSetting(translator, 0, layerArgument), new FuelSetting(translator, 0, layerArgument), modifiesFrame, stripLits);
-      }
-
-      public ExpressionTranslator WithCustomFuelSetting(CustomFuelSettings customSettings) {
-        // Use the existing layers but with some per-function customizations
-        Contract.Requires(customSettings != null);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, null, layerInterCluster.WithContext(customSettings), layerIntraCluster.WithContext(customSettings), modifiesFrame, stripLits);
-      }
-
-      public ExpressionTranslator ReplaceLayer(Bpl.Expr layerArgument) {
-        // different layer with same fuel amount.
-        Contract.Requires(layerArgument != null);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, applyLimited_CurrentFunction, layerInterCluster.WithLayer(layerArgument), layerIntraCluster.WithLayer(layerArgument), modifiesFrame, stripLits);
-       }
-
-      public ExpressionTranslator WithNoLits() {
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, applyLimited_CurrentFunction, layerInterCluster, layerIntraCluster, modifiesFrame, true);
-      }
-
-      public ExpressionTranslator LimitedFunctions(Function applyLimited_CurrentFunction, Bpl.Expr layerArgument) {
-        Contract.Requires(applyLimited_CurrentFunction != null);
-        Contract.Requires(layerArgument != null);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, applyLimited_CurrentFunction, /* layerArgument */ layerInterCluster, new FuelSetting(translator, 0, layerArgument), modifiesFrame, stripLits);
-      }
-
-      public ExpressionTranslator LayerOffset(int offset) {
-        Contract.Requires(0 <= offset);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, applyLimited_CurrentFunction, layerInterCluster.Offset(offset), layerIntraCluster, modifiesFrame, stripLits);
-      }
-
-      public ExpressionTranslator DecreaseFuel(int offset) {
-        Contract.Requires(0 <= offset);
-        Contract.Ensures(Contract.Result<ExpressionTranslator>() != null);
-
-        return CloneExpressionTranslator(this, translator, predef, HeapExpr, This, applyLimited_CurrentFunction, layerInterCluster.Decrease(offset), layerIntraCluster, modifiesFrame, stripLits);
-      }
-
-      private static ExpressionTranslator CloneExpressionTranslator(ExpressionTranslator orig,
-        Translator translator, PredefinedDecls predef, Bpl.Expr heap, string thisVar,
-        Function applyLimited_CurrentFunction, FuelSetting layerInterCluster, FuelSetting layerIntraCluster, string modifiesFrame, bool stripLits) {
-        var et = new ExpressionTranslator(translator, predef, heap, thisVar, applyLimited_CurrentFunction, layerInterCluster, layerIntraCluster, modifiesFrame, stripLits);
-        if (orig.oldEtran != null) {
-          var etOld = new ExpressionTranslator(translator, predef, orig.Old.HeapExpr, thisVar, applyLimited_CurrentFunction, layerInterCluster, layerIntraCluster, modifiesFrame, stripLits);
-          etOld.oldEtran = etOld;
-          et.oldEtran = etOld;
-        }
-        return et;
-      }
-
-      public Bpl.IdentifierExpr TheFrame(IToken tok)
-      {
-        Contract.Requires(tok != null);
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>() != null);
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>().Type != null);
-
-        Bpl.TypeVariable alpha = new Bpl.TypeVariable(tok, "beta");
-        Bpl.Type fieldAlpha = predef.FieldName(tok, alpha);
-        Bpl.Type ty = new Bpl.MapType(tok, new List<TypeVariable> { alpha }, new List<Bpl.Type> { predef.RefType, fieldAlpha }, Bpl.Type.Bool);
-        return new Bpl.IdentifierExpr(tok, this.modifiesFrame, ty);
-      }
-
-      public Bpl.IdentifierExpr Tick() {
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>() != null);
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>().Type != null);
-        return new Bpl.IdentifierExpr(Token.NoToken, "$Tick", predef.TickType);
-      }
-
-      public Bpl.IdentifierExpr ArbitraryBoxValue() {
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>() != null);
-        return new Bpl.IdentifierExpr(Token.NoToken, "$ArbitraryBoxValue", predef.BoxType);
-      }
-      public Bpl.Expr ArbitraryValue(Type type) {
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-        var bx = ArbitraryBoxValue();
-        if (!ModeledAsBoxType(type)) {
-          return translator.FunctionCall(Token.NoToken, BuiltinFunction.Unbox, translator.TrType(type), bx);
-        } else {
-          return bx;
-        }
-      }
-
-      public Bpl.IdentifierExpr FunctionContextHeight() {
-        Contract.Ensures(Contract.Result<Bpl.IdentifierExpr>().Type != null);
-        return new Bpl.IdentifierExpr(Token.NoToken, "$FunctionContextHeight", Bpl.Type.Int);
-      }
-
-      public Bpl.Expr HeightContext(ICallable m)
-      {
-        Contract.Requires(m != null);
-        // free requires fh == FunctionContextHeight;
-        var module = m.EnclosingModule;
-        Bpl.Expr context =
-          Bpl.Expr.Eq(Bpl.Expr.Literal(module.CallGraph.GetSCCRepresentativeId(m)), FunctionContextHeight());
-        return context;
-      }
-
-      public Expression GetSubstitutedBody(LetExpr e)
-      {
-        Contract.Requires(e != null);
-        Contract.Requires(e.Exact);
-        Contract.Assert(e.LHSs.Count == e.RHSs.Count);  // checked by resolution
-        var substMap = new Dictionary<IVariable, Expression>();
-        for (int i = 0; i < e.LHSs.Count; i++) {
-          translator.AddCasePatternVarSubstitutions(e.LHSs[i], TrExpr(e.RHSs[i]), substMap);
-        }
-        return Translator.Substitute(e.Body, null, substMap);
-      }
-
-      public Expr MaybeLit(Expr expr, Bpl.Type type) {
-        return stripLits ? expr : translator.Lit(expr, type);
-      }
-
-      public Expr MaybeLit(Expr expr) {
-        return stripLits ? expr : translator.Lit(expr);
-      }
-
-      /// <summary>
-      /// Translates Dafny expression "expr" into a Boogie expression.  If the type of "expr" can be a boolean, then the
-      /// token (source location) of the resulting expression is filled in (it wouldn't hurt if the token were always
-      /// filled in, but it is really necessary for anything that may show up in a Boogie assert, since that location may
-      /// then show up in an error message).
-      /// </summary>
-      public Bpl.Expr TrExpr(Expression expr) {
-        Contract.Requires(expr != null);
-        Contract.Requires(predef != null);
-
-        if (expr is LiteralExpr) {
-          LiteralExpr e = (LiteralExpr)expr;
-          if (e.Value == null) {
-            return predef.Null;
-          } else if (e.Value is bool) {
-            return MaybeLit(new Bpl.LiteralExpr(e.tok, (bool)e.Value));
-          } else if (e is CharLiteralExpr) {
-            // we expect e.Value to be a string representing exactly one char
-            Bpl.Expr rawElement = null;  // assignment to please compiler's definite assignment rule
-            foreach (char ch in Util.UnescapedCharacters((string)e.Value, false)) {
-              Contract.Assert(rawElement == null);  // we should get here only once
-              rawElement = translator.FunctionCall(expr.tok, BuiltinFunction.CharFromInt, null, Bpl.Expr.Literal((int)ch));
-            }
-            Contract.Assert(rawElement != null);  // there should have been an iteration of the loop above
-            return MaybeLit(rawElement, predef.CharType);
-          } else if (e is StringLiteralExpr) {
-            var str = (StringLiteralExpr)e;
-            Bpl.Expr seq = translator.FunctionCall(expr.tok, BuiltinFunction.SeqEmpty, predef.BoxType);
-            foreach (char ch in Util.UnescapedCharacters((string)e.Value, str.IsVerbatim)) {
-              var rawElement = translator.FunctionCall(expr.tok, BuiltinFunction.CharFromInt, null, Bpl.Expr.Literal((int)ch));
-              Bpl.Expr elt = BoxIfNecessary(expr.tok, rawElement, Type.Char);
-              seq = translator.FunctionCall(expr.tok, BuiltinFunction.SeqBuild, predef.BoxType, seq, elt);
-            }
-            return MaybeLit(seq, translator.TrType(new SeqType(Type.Char)));
-          } else if (e.Value is BigInteger) {
-            var n = Microsoft.BaseTypes.BigNum.FromBigInt((BigInteger)e.Value);
-            if (e.Type is BitvectorType) {
-              return MaybeLit(translator.BplBvLiteralExpr(e.tok, n, (BitvectorType)e.Type));
-            } else if (e.Type.IsBigOrdinalType) {
-              var fromNat = translator.FunctionCall(expr.tok, "ORD#FromNat", predef.BigOrdinalType, Bpl.Expr.Literal(n));
-              return MaybeLit(fromNat, predef.BigOrdinalType);
-            } else {
-              return MaybeLit(Bpl.Expr.Literal(n));
-            }
-          } else if (e.Value is BaseTypes.BigDec) {
-            return MaybeLit(Bpl.Expr.Literal((BaseTypes.BigDec)e.Value));
-          } else {
-            Contract.Assert(false); throw new cce.UnreachableException();  // unexpected literal
-          }
-
-        } else if (expr is ThisExpr) {
-          return new Bpl.IdentifierExpr(expr.tok, This, translator.TrType(expr.Type));
-
-        } else if (expr is IdentifierExpr) {
-          IdentifierExpr e = (IdentifierExpr)expr;
-          Contract.Assert(e.Var != null);
-          return translator.TrVar(expr.tok, e.Var);
-
-        } else if (expr is BoogieWrapper) {
-          var e = (BoogieWrapper)expr;
-          return e.Expr;
-
-
-        } else if (expr is BoogieFunctionCall) {
-          var e = (BoogieFunctionCall)expr;
-          var id = new Bpl.IdentifierExpr(e.tok, e.FunctionName, translator.TrType(e.Type));
-          var args = new List<Bpl.Expr>();
-          foreach (var arg in e.TyArgs) {
-            args.Add(translator.TypeToTy(arg));
-          }
-          if (e.UsesHeap) {
-            args.Add(HeapExpr);
-          }
-          if (e.UsesOldHeap) {
-            args.Add(Old.HeapExpr);
-          }
-          foreach (var heapAtLabel in e.HeapAtLabels) {
-            Bpl.Expr ve;
-            var bv = BplBoundVar("$Heap_at_" + heapAtLabel.AssignUniqueId(translator.CurrentIdGenerator), translator.predef.HeapType, out ve);
-            args.Add(ve);
-          }
-          foreach (var arg in e.Args) {
-            args.Add(TrExpr(arg));
-          }
-          return new Bpl.NAryExpr(e.tok, new Bpl.FunctionCall(id), args);
-
-        } else if (expr is SetDisplayExpr) {
-          SetDisplayExpr e = (SetDisplayExpr)expr;
-          Bpl.Expr s = translator.FunctionCall(expr.tok, e.Finite ? BuiltinFunction.SetEmpty : BuiltinFunction.ISetEmpty, predef.BoxType);
-          var isLit = true;
-          foreach (Expression ee in e.Elements) {
-            var rawElement = TrExpr(ee);
-            isLit = isLit && translator.IsLit(rawElement);
-            Bpl.Expr ss = BoxIfNecessary(expr.tok, rawElement, cce.NonNull(ee.Type));
-            s = translator.FunctionCall(expr.tok, e.Finite ? BuiltinFunction.SetUnionOne : BuiltinFunction.ISetUnionOne, predef.BoxType, s, ss);
-          }
-          if (isLit) {
-            // Lit-lifting: All elements are lit, so the set is Lit too
-            s = MaybeLit(s, predef.BoxType);
-          }
-          return s;
-
-        } else if (expr is MultiSetDisplayExpr) {
-          MultiSetDisplayExpr e = (MultiSetDisplayExpr)expr;
-          Bpl.Expr s = translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetEmpty, predef.BoxType);
-          var isLit = true;
-          foreach (Expression ee in e.Elements) {
-            var rawElement = TrExpr(ee);
-            isLit = isLit && translator.IsLit(rawElement);
-            Bpl.Expr ss = BoxIfNecessary(expr.tok, rawElement, cce.NonNull(ee.Type));
-            s = translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetUnionOne, predef.BoxType, s, ss);
-          }
-          if (isLit) {
-            // Lit-lifting: All elements are lit, so the multiset is Lit too
-            s = MaybeLit(s, predef.BoxType);
-          }
-          return s;
-
-        } else if (expr is SeqDisplayExpr) {
-          SeqDisplayExpr e = (SeqDisplayExpr)expr;
-          // Note: a LiteralExpr(string) is really another kind of SeqDisplayExpr
-          Bpl.Expr s = translator.FunctionCall(expr.tok, BuiltinFunction.SeqEmpty, predef.BoxType);
-          var isLit = true;
-          foreach (Expression ee in e.Elements) {
-            var rawElement = TrExpr(ee);
-            isLit = isLit && translator.IsLit(rawElement);
-            Bpl.Expr elt = BoxIfNecessary(expr.tok, rawElement, ee.Type);
-            s = translator.FunctionCall(expr.tok, BuiltinFunction.SeqBuild, predef.BoxType, s, elt);
-          }
-          if (isLit) {
-            // Lit-lifting: All elements are lit, so the sequence is Lit too
-            s = MaybeLit(s, predef.BoxType);
-          }
-          return s;
-
-        } else if (expr is MapDisplayExpr) {
-          MapDisplayExpr e = (MapDisplayExpr)expr;
-          Bpl.Type maptype = predef.MapType(expr.tok, e.Finite, predef.BoxType, predef.BoxType);
-          Bpl.Expr s = translator.FunctionCall(expr.tok, e.Finite ? BuiltinFunction.MapEmpty : BuiltinFunction.IMapEmpty, predef.BoxType);
-          var isLit = true;
-          foreach (ExpressionPair p in e.Elements) {
-            var rawA = TrExpr(p.A);
-            var rawB = TrExpr(p.B);
-            isLit = isLit && translator.IsLit(rawA) && translator.IsLit(rawB);
-            Bpl.Expr elt = BoxIfNecessary(expr.tok, rawA, cce.NonNull(p.A.Type));
-            Bpl.Expr elt2 = BoxIfNecessary(expr.tok, rawB, cce.NonNull(p.B.Type));
-            s = translator.FunctionCall(expr.tok, e.Finite ? "Map#Build" : "IMap#Build", maptype, s, elt, elt2);
-          }
-          if (isLit) {
-            // Lit-lifting: All keys and values are lit, so the map is Lit too
-            s = MaybeLit(s, predef.BoxType);
-          }
-          return s;
-
-        } else if (expr is MemberSelectExpr) {
-          var e = (MemberSelectExpr)expr;
-          return e.MemberSelectCase(
-            field => {
-              var useSurrogateLocal = translator.inBodyInitContext && Expression.AsThis(e.Obj) != null && !field.IsInstanceIndependentConstant;
-              if (useSurrogateLocal) {
-                return new Bpl.IdentifierExpr(expr.tok, translator.SurrogateName(field), translator.TrType(field.Type));
-              } else if (field is ConstantField) {
-                var typeMap = e.TypeArgumentSubstitutionsWithParents();
-                var args = GetTypeParams(field.EnclosingClass).ConvertAll(tp => translator.TypeToTy(typeMap[tp]));
-                Bpl.Expr result;
-                if (field.IsStatic) {
-                  result = new Bpl.NAryExpr(expr.tok, new Bpl.FunctionCall(translator.GetReadonlyField(field)), args);
-                } else {
-                  Bpl.Expr obj = TrExpr(e.Obj);
-                  args.Add(obj);
-                  result = new Bpl.NAryExpr(expr.tok, new Bpl.FunctionCall(translator.GetReadonlyField(field)), args);
-                }
-                result = translator.CondApplyUnbox(expr.tok, result, field.Type, expr.Type);
-                return result;
-              } else {
-                Bpl.Expr obj = TrExpr(e.Obj);
-                Bpl.Expr result;
-                if (field.IsMutable) {
-                  result = ReadHeap(expr.tok, HeapExpr, obj, new Bpl.IdentifierExpr(expr.tok, translator.GetField(field)));
-                  return translator.CondApplyUnbox(expr.tok, result, field.Type, expr.Type);
-                } else {
-                  result = new Bpl.NAryExpr(expr.tok, new Bpl.FunctionCall(translator.GetReadonlyField(field)),
-                    new List<Bpl.Expr> { obj });
-                  result = translator.CondApplyUnbox(expr.tok, result, field.Type, expr.Type);
-                  if (translator.IsLit(obj)) {
-                    result = MaybeLit(result, translator.TrType(expr.Type));
-                  }
-                  return result;
-                }
-              }
-            },
-            fn => {
-              var typeMap = e.TypeArgumentSubstitutionsWithParents();
-              var args = GetTypeParams(fn).ConvertAll(tp => translator.TypeToTy(typeMap[tp]));
-              if (fn.IsFuelAware()) {
-                args.Add(this.layerInterCluster.GetFunctionFuel(fn));
-              }
-              if (fn is TwoStateFunction) {
-                args.Add(Old.HeapExpr);
-              }
-              if (!fn.IsStatic) {
-                args.Add(/* translator.BoxIfUnboxed */(TrExpr(e.Obj)/*, e.Type */));
-              }
-              return translator.FunctionCall(e.tok, translator.FunctionHandle(fn), predef.HandleType, args);
-            });
-        } else if (expr is SeqSelectExpr) {
-          SeqSelectExpr e = (SeqSelectExpr)expr;
-          Bpl.Expr seq = TrExpr(e.Seq);
-          var seqType = e.Seq.Type.NormalizeExpand();
-          Type elmtType = null;
-          Type domainType = null;
-          Contract.Assert(seqType != null);  // the expression has been successfully resolved
-          if (seqType.IsArrayType) {
-            domainType = Type.Int;
-            elmtType = UserDefinedType.ArrayElementType(seqType);
-          } else if (seqType is SeqType) {
-            domainType = Type.Int;
-            elmtType = ((SeqType)seqType).Arg;
-          } else if (seqType is MapType) {
-            domainType = ((MapType)seqType).Domain;
-            elmtType = ((MapType)seqType).Range;
-          } else if (seqType is MultiSetType) {
-            domainType = ((MultiSetType)seqType).Arg;
-            elmtType = Type.Int;
-          } else { Contract.Assert(false); }
-          Bpl.Type elType = translator.TrType(elmtType);
-          Bpl.Type dType = translator.TrType(domainType);
-          Bpl.Expr e0 = e.E0 == null ? null : TrExpr(e.E0);
-          if (e0 != null && e.E0.Type.IsBitVectorType) {
-            e0 = translator.ConvertExpression(e.E0.tok, e0, e.E0.Type, Type.Int);
-          }
-          Bpl.Expr e1 = e.E1 == null ? null : TrExpr(e.E1);
-          if (e1 != null && e.E1.Type.IsBitVectorType) {
-            e1 = translator.ConvertExpression(e.E1.tok, e1, e.E1.Type, Type.Int);
-          }
-          if (e.SelectOne) {
-            Contract.Assert(e1 == null);
-            Bpl.Expr x;
-            if (seqType.IsArrayType) {
-              Bpl.Expr fieldName = translator.FunctionCall(expr.tok, BuiltinFunction.IndexField, null, e0);
-              x = ReadHeap(expr.tok, HeapExpr, TrExpr(e.Seq), fieldName);
-            } else if (seqType is SeqType) {
-              x = translator.FunctionCall(expr.tok, BuiltinFunction.SeqIndex, predef.BoxType, seq, e0);
-            } else if (seqType is MapType) {
-              bool finite = ((MapType)seqType).Finite;
-              var f = finite ? BuiltinFunction.MapElements : BuiltinFunction.IMapElements;
-              x = translator.FunctionCall(expr.tok, f, predef.MapType(e.tok, finite, predef.BoxType, predef.BoxType), seq);
-              x = Bpl.Expr.Select(x, BoxIfNecessary(e.tok, e0, domainType));
-            } else if (seqType is MultiSetType) {
-              x = Bpl.Expr.SelectTok(expr.tok, TrExpr(e.Seq), BoxIfNecessary(expr.tok, e0, domainType));
-            } else { Contract.Assert(false); x = null; }
-            if (!ModeledAsBoxType(elmtType) && !(seqType is MultiSetType)) {
-              x = translator.FunctionCall(expr.tok, BuiltinFunction.Unbox, elType, x);
-            }
-            return x;
-          } else {
-            if (seqType.IsArrayType) {
-              seq = translator.FunctionCall(expr.tok, BuiltinFunction.SeqFromArray, elType, HeapExpr, seq);
-            }
-            var isLit = translator.IsLit(seq);
-            if (e1 != null) {
-              isLit = isLit && translator.IsLit(e1);
-              seq = translator.FunctionCall(expr.tok, BuiltinFunction.SeqTake, elType, seq, e1);
-            }
-            if (e0 != null) {
-              isLit = isLit && translator.IsLit(e0);
-              seq = translator.FunctionCall(expr.tok, BuiltinFunction.SeqDrop, elType, seq, e0);
-            }
-            // if e0 == null && e1 == null, then we have the identity operation seq[..] == seq;
-            if (isLit && (e0 != null || e1 != null)) {
-              // Lit-lift the expression
-              seq = MaybeLit(seq, translator.TrType(expr.Type));
-            }
-            return seq;
-          }
-
-        } else if (expr is SeqUpdateExpr) {
-          SeqUpdateExpr e = (SeqUpdateExpr)expr;
-          if (e.ResolvedUpdateExpr != null)
-          {
-            return TrExpr(e.ResolvedUpdateExpr);
-          }
-          else
-          {
-            Bpl.Expr seq = TrExpr(e.Seq);
-            var seqType = e.Seq.Type.NormalizeExpand();
-            if (seqType is SeqType)
-            {
-              Type elmtType = cce.NonNull((SeqType)seqType).Arg;
-              Bpl.Expr index = TrExpr(e.Index);
-              index = translator.ConvertExpression(e.Index.tok, index, e.Index.Type, Type.Int);
-              Bpl.Expr val = BoxIfNecessary(expr.tok, TrExpr(e.Value), elmtType);
-              return translator.FunctionCall(expr.tok, BuiltinFunction.SeqUpdate, predef.BoxType, seq, index, val);
-            }
-            else if (seqType is MapType)
-            {
-              MapType mt = (MapType)seqType;
-              Bpl.Type maptype = predef.MapType(expr.tok, mt.Finite, predef.BoxType, predef.BoxType);
-              Bpl.Expr index = BoxIfNecessary(expr.tok, TrExpr(e.Index), mt.Domain);
-              Bpl.Expr val = BoxIfNecessary(expr.tok, TrExpr(e.Value), mt.Range);
-              return translator.FunctionCall(expr.tok, mt.Finite ? "Map#Build" : "IMap#Build", maptype, seq, index, val);
-            }
-            else if (seqType is MultiSetType)
-            {
-              Type elmtType = cce.NonNull((MultiSetType)seqType).Arg;
-              Bpl.Expr index = BoxIfNecessary(expr.tok, TrExpr(e.Index), elmtType);
-              Bpl.Expr val = TrExpr(e.Value);
-              return Bpl.Expr.StoreTok(expr.tok, seq, index, val);
-            }
-            else
-            {
-              Contract.Assert(false);
-              throw new cce.UnreachableException();
-            }
-          }
-
-        } else if (expr is MultiSelectExpr) {
-          MultiSelectExpr e = (MultiSelectExpr)expr;
-          Type elmtType = UserDefinedType.ArrayElementType(e.Array.Type);;
-          Bpl.Type elType = translator.TrType(elmtType);
-
-          Bpl.Expr fieldName = GetArrayIndexFieldName(expr.tok, e.Indices);
-          Bpl.Expr x = ReadHeap(expr.tok, HeapExpr, TrExpr(e.Array), fieldName);
-          if (!ModeledAsBoxType(elmtType)) {
-            x = translator.FunctionCall(expr.tok, BuiltinFunction.Unbox, elType, x);
-          }
-          return x;
-
-        } else if (expr is ApplyExpr) {
-          ApplyExpr e = (ApplyExpr)expr;
-          int arity = e.Args.Count;
-          var tt = e.Function.Type.AsArrowType;
-          Contract.Assert(tt != null);
-          Contract.Assert(tt.Arity == arity);
-
-          {
-            // optimisation: if this could have just as well been a FunctionCallExpr, call it as such!
-            var con = e.Function as ConcreteSyntaxExpression;
-            var recv = con == null ? e.Function : con.Resolved;
-            var mem = recv as MemberSelectExpr;
-            var fn = mem == null ? null : mem.Member as Function;
-            if (fn != null) {
-              return TrExpr(new FunctionCallExpr(e.tok, fn.Name, mem.Obj, e.tok, e.Args) {
-                Function = fn,
-                Type = e.Type,
-                TypeApplication_AtEnclosingClass = mem.TypeApplication_AtEnclosingClass,
-                TypeApplication_JustFunction = mem.TypeApplication_JustMember
-              });
-            }
-          }
-
-          Func<Expression, Bpl.Expr> TrArg = arg => translator.BoxIfUnboxed(TrExpr(arg), arg.Type);
-
-          var applied = translator.FunctionCall(expr.tok, Translator.Apply(arity), predef.BoxType,
-            Concat(Map(tt.TypeArgs,translator.TypeToTy),
-            Cons(HeapExpr, Cons(TrExpr(e.Function), e.Args.ConvertAll(arg => TrArg(arg))))));
-
-          return translator.UnboxIfBoxed(applied, tt.Result);
-
-        } else if (expr is FunctionCallExpr) {
-          FunctionCallExpr e = (FunctionCallExpr)expr;
-          if (e.Function is SpecialFunction) {
-            return TrExprSpecialFunctionCall(e);
-          } else {
-            Bpl.Expr layerArgument;
-            var etran = this;
-            if (e.Function.ContainsQuantifier && translator.stmtContext == StmtType.ASSUME && translator.adjustFuelForExists) {
-              // we need to increase fuel functions that contain quantifier expr in the assume context.
-              etran = etran.LayerOffset(1);
-              translator.adjustFuelForExists = false;
-            }
-            if (e.Function.IsFuelAware()) {
-              Statistics_CustomLayerFunctionCount++;
-              ModuleDefinition module = e.Function.EnclosingClass.EnclosingModuleDefinition;
-              if (etran.applyLimited_CurrentFunction != null &&
-                etran.layerIntraCluster != null &&
-                ModuleDefinition.InSameSCC(e.Function, applyLimited_CurrentFunction)) {
-                layerArgument = etran.layerIntraCluster.GetFunctionFuel(e.Function);
-              } else {
-                layerArgument = etran.layerInterCluster.GetFunctionFuel(e.Function);
-              }
-            } else {
-              layerArgument = null;
-            }
-
-            var ty = translator.TrType(e.Type);
-            var id = new Bpl.IdentifierExpr(e.tok, e.Function.FullSanitizedName, ty);
-
-            bool argsAreLit;
-            var args = FunctionInvocationArguments(e, layerArgument, false, out argsAreLit);
-            Expr result = new Bpl.NAryExpr(e.tok, new Bpl.FunctionCall(id), args);
-            result = translator.CondApplyUnbox(e.tok, result, e.Function.ResultType, e.Type);
-
-            bool callIsLit = argsAreLit
-              && Translator.FunctionBodyIsAvailable(e.Function, translator.currentModule, translator.currentScope, true)
-              && !e.Function.Reads.Any(); // Function could depend on external values
-            if (callIsLit) {
-              result = MaybeLit(result, ty);
-            }
-
-            return result;
-          }
-        } else if (expr is DatatypeValue) {
-          DatatypeValue dtv = (DatatypeValue)expr;
-          Contract.Assert(dtv.Ctor != null);  // since dtv has been successfully resolved
-          List<Bpl.Expr> args = new List<Bpl.Expr>();
-
-          bool argsAreLit = true;
-          for (int i = 0; i < dtv.Arguments.Count; i++) {
-            Expression arg = dtv.Arguments[i];
-            Type t = dtv.Ctor.Formals[i].Type;
-            var bArg = TrExpr(arg);
-            argsAreLit = argsAreLit && translator.IsLit(bArg);
-            args.Add(translator.CondApplyBox(expr.tok, bArg, cce.NonNull(arg.Type), t));
-          }
-          Bpl.IdentifierExpr id = new Bpl.IdentifierExpr(dtv.tok, dtv.Ctor.FullName, predef.DatatypeType);
-          Bpl.Expr ret = new Bpl.NAryExpr(dtv.tok, new Bpl.FunctionCall(id), args);
-          if (argsAreLit) {
-            // If all arguments are Lit, so is the whole expression
-            ret = MaybeLit(ret, predef.DatatypeType);
-          }
-          return ret;
-
-        } else if (expr is SeqConstructionExpr) {
-          var e = (SeqConstructionExpr)expr;
-          var eType = e.Type.AsSeqType.Arg.NormalizeExpand();
-          return translator.FunctionCall(expr.tok, "Seq#Create", predef.SeqType(e.tok, predef.BoxType), translator.TypeToTy(eType), HeapExpr, TrExpr(e.N), TrExpr(e.Initializer));
-
-        } else if (expr is MultiSetFormingExpr) {
-          MultiSetFormingExpr e = (MultiSetFormingExpr)expr;
-          var eType = e.E.Type.NormalizeExpand();
-          if (eType is SetType) {
-            return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetFromSet, translator.TrType(cce.NonNull((SetType)eType).Arg), TrExpr(e.E));
-          } else if (eType is SeqType) {
-            return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetFromSeq, translator.TrType(cce.NonNull((SeqType)eType).Arg), TrExpr(e.E));
-          } else {
-            Contract.Assert(false); throw new cce.UnreachableException();
-          }
-
-        } else if (expr is OldExpr) {
-          var e = (OldExpr)expr;
-          return OldAt(e.AtLabel).TrExpr(e.E);
-
-        } else if (expr is UnchangedExpr) {
-          var e = (UnchangedExpr)expr;
-          return translator.FrameCondition(e.tok, e.Frame, false, Resolver.FrameExpressionUse.Unchanged, OldAt(e.AtLabel), this, this, true);
-
-        } else if (expr is UnaryOpExpr) {
-          var e = (UnaryOpExpr)expr;
-          Bpl.Expr arg = TrExpr(e.E);
-          switch (e.Op) {
-            case UnaryOpExpr.Opcode.Lit:
-              return MaybeLit(arg);
-            case UnaryOpExpr.Opcode.Not:
-              if (expr.Type.IsBitVectorType) {
-                var bvWidth = expr.Type.AsBitVectorType.Width;
-                var bvType = translator.BplBvType(bvWidth);
-                Bpl.Expr r = translator.FunctionCall(expr.tok, "not_bv" + bvWidth, bvType, arg);
-                if (translator.IsLit(arg)) {
-                  r = MaybeLit(r, bvType);
-                }
-                return r;
-              } else {
-                return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, arg);
-              }
-            case UnaryOpExpr.Opcode.Cardinality:
-              var eType = e.E.Type.NormalizeExpand();
-              if (eType is SeqType) {
-                return translator.FunctionCall(expr.tok, BuiltinFunction.SeqLength, null, arg);
-              } else if (eType is SetType && ((SetType)eType).Finite) {
-                return translator.FunctionCall(expr.tok, BuiltinFunction.SetCard, null, arg);
-              } else if (eType is MultiSetType) {
-                return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetCard, null, arg);
-              } else if (eType is MapType && ((MapType)eType).Finite) {
-                return translator.FunctionCall(expr.tok, BuiltinFunction.MapCard, null, arg);
-              } else {
-                Contract.Assert(false); throw new cce.UnreachableException();  // unexpected sized type
-              }
-            case UnaryOpExpr.Opcode.Fresh:
-              var freshLabel = ((FreshExpr)e).AtLabel;
-              var eeType = e.E.Type.NormalizeExpand();
-              if (eeType is SetType) {
-                // generate:  (forall $o: ref :: { X[Box($o)] } X[Box($o)] ==> $o != null && !old($Heap)[$o,alloc])
-                // OR, if X[Box($o)] is rewritten into smaller parts, use the less good trigger old($Heap)[$o,alloc]
-                Bpl.Variable oVar = new Bpl.BoundVariable(expr.tok, new Bpl.TypedIdent(expr.tok, "$o", predef.RefType));
-                Bpl.Expr o = new Bpl.IdentifierExpr(expr.tok, oVar);
-                Bpl.Expr oNotNull = Bpl.Expr.Neq(o, predef.Null);
-                bool performedInSetRewrite;
-                Bpl.Expr oInSet = TrInSet(expr.tok, o, e.E, ((SetType)eeType).Arg, true, out performedInSetRewrite);
-                Bpl.Expr oNotFresh = OldAt(freshLabel).IsAlloced(expr.tok, o);
-                Bpl.Expr oIsFresh = Bpl.Expr.Not(oNotFresh);
-                Bpl.Expr body = Bpl.Expr.Imp(oInSet, Bpl.Expr.And(oNotNull, oIsFresh));
-                var trigger = BplTrigger(performedInSetRewrite ? oNotFresh : oInSet);
-                return new Bpl.ForallExpr(expr.tok, new List<Variable> { oVar }, trigger, body);
-              } else if (eeType is SeqType) {
-                // generate:  (forall $i: int :: 0 <= $i && $i < Seq#Length(X) ==> Unbox(Seq#Index(X,$i)) != null && !old($Heap)[Unbox(Seq#Index(X,$i)),alloc])
-                Bpl.Variable iVar = new Bpl.BoundVariable(expr.tok, new Bpl.TypedIdent(expr.tok, "$i", Bpl.Type.Int));
-                Bpl.Expr i = new Bpl.IdentifierExpr(expr.tok, iVar);
-                Bpl.Expr iBounds = translator.InSeqRange(expr.tok, i, Type.Int, TrExpr(e.E), true, null, false);
-                Bpl.Expr XsubI = translator.FunctionCall(expr.tok, BuiltinFunction.SeqIndex, predef.RefType, TrExpr(e.E), i);
-                XsubI = translator.FunctionCall(expr.tok, BuiltinFunction.Unbox, predef.RefType, XsubI);
-                Bpl.Expr oNotFresh = OldAt(freshLabel).IsAlloced(expr.tok, XsubI);
-                Bpl.Expr oIsFresh = Bpl.Expr.Not(oNotFresh);
-                Bpl.Expr xsubiNotNull = Bpl.Expr.Neq(XsubI, predef.Null);
-                Bpl.Expr body = Bpl.Expr.Imp(iBounds, Bpl.Expr.And(xsubiNotNull, oIsFresh));
-                //TRIGGERS: Does this make sense? dafny0\SmallTests
-                // BROKEN // NEW_TRIGGER
-                //TRIG (forall $i: int :: 0 <= $i && $i < Seq#Length(Q#0) && $Unbox(Seq#Index(Q#0, $i)): ref != null ==> !read(old($Heap), $Unbox(Seq#Index(Q#0, $i)): ref, alloc))
-                return new Bpl.ForallExpr(expr.tok, new List<Variable> { iVar }, body);
-              } else {
-                // generate:  x != null && !old($Heap)[x]
-                Bpl.Expr oNull = Bpl.Expr.Neq(TrExpr(e.E), predef.Null);
-                Bpl.Expr oIsFresh = Bpl.Expr.Not(OldAt(freshLabel).IsAlloced(expr.tok, TrExpr(e.E)));
-                return Bpl.Expr.Binary(expr.tok, BinaryOperator.Opcode.And, oNull, oIsFresh);
-              }
-            case UnaryOpExpr.Opcode.Allocated: {
-                var aType = e.E.Type.NormalizeExpand();
-                return translator.MkIsAlloc(TrExpr(e.E), aType, HeapExpr);
-              }
-            default:
-              Contract.Assert(false); throw new cce.UnreachableException();  // unexpected unary expression
-          }
-
-        } else if (expr is ConversionExpr) {
-          var e = (ConversionExpr)expr;
-          return translator.ConvertExpression(e.tok, TrExpr(e.E), e.E.Type, e.ToType);
-
-        } else if (expr is TypeTestExpr) {
-          var e = (TypeTestExpr)expr;
-          return translator.GetSubrangeCheck(TrExpr(e.E), e.E.Type, e.ToType, out var _) ?? Bpl.Expr.True;
-
-        } else if (expr is BinaryExpr) {
-          BinaryExpr e = (BinaryExpr)expr;
-          bool isReal = e.E0.Type.IsNumericBased(Type.NumericPersuasion.Real);
-          int bvWidth = e.E0.Type.IsBitVectorType ? e.E0.Type.AsBitVectorType.Width : -1;  // -1 indicates "not a bitvector type"
-          Bpl.Expr e0 = TrExpr(e.E0);
-          if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.InSet) {
-            bool pr;
-            return TrInSet(expr.tok, e0, e.E1, cce.NonNull(e.E0.Type), false, out pr);  // let TrInSet translate e.E1
-          } else if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.NotInSet) {
-            bool pr;
-            Bpl.Expr arg = TrInSet(expr.tok, e0, e.E1, cce.NonNull(e.E0.Type), false, out pr);  // let TrInSet translate e.E1
-            return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, arg);
-          } else if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.InMultiSet) {
-            return TrInMultiSet(expr.tok, e0, e.E1, cce.NonNull(e.E0.Type), false); // let TrInMultiSet translate e.E1
-          } else if (e.ResolvedOp == BinaryExpr.ResolvedOpcode.NotInMultiSet) {
-            Bpl.Expr arg = TrInMultiSet(expr.tok, e0, e.E1, cce.NonNull(e.E0.Type), false);  // let TrInMultiSet translate e.E1
-            return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, arg);
-          }
-          Bpl.Expr e1 = TrExpr(e.E1);
-          BinaryOperator.Opcode bOpcode;
-          Bpl.Type typ;
-          var oe0 = e0;
-          var oe1 = e1;
-          var lit0 = translator.GetLit(e0);
-          var lit1 = translator.GetLit(e1);
-          bool liftLit = translator.IsLit(e0) && translator.IsLit(e1);
-          // NOTE(namin): We usually avoid keeping literals, because their presence might mess up triggers that do not expect them.
-          //              Still for equality-related operations, it's useful to keep them instead of lifting them, so that they can be propagated.
-          bool keepLits = false;
-          if (lit0 != null) {
-            e0 = lit0;
-          }
-          if (lit1 != null) {
-            e1 = lit1;
-          }
-          switch (e.ResolvedOp) {
-            case BinaryExpr.ResolvedOpcode.Iff:
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.Iff; break;
-            case BinaryExpr.ResolvedOpcode.Imp:
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.Imp; break;
-            case BinaryExpr.ResolvedOpcode.And:
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.And; break;
-            case BinaryExpr.ResolvedOpcode.Or:
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.Or; break;
-
-            case BinaryExpr.ResolvedOpcode.EqCommon:
-              keepLits = true;
-              var cot = e.E0.Type.AsCoDatatype;
-              if (cot != null) {
-                var e0args = e.E0.Type.NormalizeExpand().TypeArgs;
-                var e1args = e.E1.Type.NormalizeExpand().TypeArgs;
-                return translator.CoEqualCall(cot, e0args, e1args, null, this.layerInterCluster.LayerN((int)FuelSetting.FuelAmount.HIGH), e0, e1, expr.tok);
-              }
-              if (e.E0.Type.IsIndDatatype) {
-                return translator.TypeSpecificEqual(expr.tok, e.E0.Type, e0, e1);
-              }
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.Eq; break;
-            case BinaryExpr.ResolvedOpcode.NeqCommon:
-              var cotx = e.E0.Type.AsCoDatatype;
-              if (cotx != null) {
-                var e0args = e.E0.Type.NormalizeExpand().TypeArgs;
-                var e1args = e.E1.Type.NormalizeExpand().TypeArgs;
-                var eq = translator.CoEqualCall(cotx, e0args, e1args, null, this.layerInterCluster.LayerN((int)FuelSetting.FuelAmount.HIGH), e0, e1, expr.tok);
-                return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, eq);
-              }
-              if (e.E0.Type.IsIndDatatype) {
-                var eq = translator.TypeSpecificEqual(expr.tok, e.E0.Type, e0, e1);
-                return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, eq);
-              }
-              typ = Bpl.Type.Bool;
-              bOpcode = BinaryOperator.Opcode.Neq; break;
-            case BinaryExpr.ResolvedOpcode.Lt:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "lt_bv" + bvWidth, Bpl.Type.Bool, e0, e1, liftLit);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                return translator.FunctionCall(expr.tok, "ORD#Less", Bpl.Type.Bool, e0, e1);
-              } else if (isReal || !DafnyOptions.O.DisableNLarith) {
-                typ = Bpl.Type.Bool;
-                bOpcode = BinaryOperator.Opcode.Lt;
-                break;
-              } else {
-                return TrToFunctionCall(expr.tok, "INTERNAL_lt_boogie", Bpl.Type.Bool, e0, e1, liftLit);
-              }
-            case BinaryExpr.ResolvedOpcode.LessThanLimit:
-              return translator.FunctionCall(expr.tok, "ORD#LessThanLimit", Bpl.Type.Bool, e0, e1);
-            case BinaryExpr.ResolvedOpcode.Le:
-              keepLits = true;
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "le_bv" + bvWidth, Bpl.Type.Bool, e0, e1, false);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                var less = translator.FunctionCall(expr.tok, "ORD#Less", Bpl.Type.Bool, e0, e1);
-                var eq = Bpl.Expr.Eq(e0, e1);
-                return BplOr(eq, less);
-              } else if (isReal || !DafnyOptions.O.DisableNLarith) {
-                typ = Bpl.Type.Bool;
-                bOpcode = BinaryOperator.Opcode.Le;
-                break;
-              } else {
-                return TrToFunctionCall(expr.tok, "INTERNAL_le_boogie", Bpl.Type.Bool, e0, e1, false);
-              }
-            case BinaryExpr.ResolvedOpcode.Ge:
-              keepLits = true;
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "ge_bv" + bvWidth, Bpl.Type.Bool, e0, e1, false);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                var less = translator.FunctionCall(expr.tok, "ORD#Less", Bpl.Type.Bool, e1, e0);
-                var eq = Bpl.Expr.Eq(e1, e0);
-                return BplOr(eq, less);
-              } else if (isReal || !DafnyOptions.O.DisableNLarith) {
-                typ = Bpl.Type.Bool;
-                bOpcode = BinaryOperator.Opcode.Ge;
-                break;
-              } else {
-                return TrToFunctionCall(expr.tok, "INTERNAL_ge_boogie", Bpl.Type.Bool, e0, e1, false);
-              }
-            case BinaryExpr.ResolvedOpcode.Gt:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "gt_bv" + bvWidth, Bpl.Type.Bool, e0, e1, liftLit);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                return translator.FunctionCall(expr.tok, "ORD#Less", Bpl.Type.Bool, e1, e0);
-              } else if (isReal || !DafnyOptions.O.DisableNLarith) {
-                typ = Bpl.Type.Bool;
-                bOpcode = BinaryOperator.Opcode.Gt;
-                break;
-              } else {
-                return TrToFunctionCall(expr.tok, "INTERNAL_gt_boogie", Bpl.Type.Bool, e0, e1, liftLit);
-              }
-
-            case BinaryExpr.ResolvedOpcode.Add:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "add_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                return TrToFunctionCall(expr.tok, "ORD#Plus", predef.BigOrdinalType, e0, e1, liftLit);
-              } else if (e.E0.Type.IsCharType) {
-                return TrToFunctionCall(expr.tok, "char#Plus", predef.CharType, e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.DisableNLarith) {
-                return TrToFunctionCall(expr.tok, "INTERNAL_add_boogie", Bpl.Type.Int, e0, e1, liftLit);
-              } else if (!isReal && (DafnyOptions.O.ArithMode == 2 || 5 <= DafnyOptions.O.ArithMode)) {
-                return TrToFunctionCall(expr.tok, "Add", Bpl.Type.Int, oe0, oe1, liftLit);
-              } else {
-                typ = isReal ? Bpl.Type.Real : Bpl.Type.Int;
-                bOpcode = BinaryOperator.Opcode.Add;
-                break;
-              }
-            case BinaryExpr.ResolvedOpcode.Sub:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "sub_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              } else if (e.E0.Type.IsBigOrdinalType) {
-                return TrToFunctionCall(expr.tok, "ORD#Minus", predef.BigOrdinalType, e0, e1, liftLit);
-              } else if (e.E0.Type.IsCharType) {
-                return TrToFunctionCall(expr.tok, "char#Minus", predef.CharType, e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.DisableNLarith) {
-                return TrToFunctionCall(expr.tok, "INTERNAL_sub_boogie", Bpl.Type.Int, e0, e1, liftLit);
-              } else if (!isReal && (DafnyOptions.O.ArithMode == 2 || 5 <= DafnyOptions.O.ArithMode)) {
-                return TrToFunctionCall(expr.tok, "Sub", Bpl.Type.Int, oe0, oe1, liftLit);
-              } else {
-                typ = isReal ? Bpl.Type.Real : Bpl.Type.Int;
-                bOpcode = BinaryOperator.Opcode.Sub;
-                break;
-              }
-            case BinaryExpr.ResolvedOpcode.Mul:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "mul_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.DisableNLarith) {
-                return TrToFunctionCall(expr.tok, "INTERNAL_mul_boogie", Bpl.Type.Int, e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.ArithMode != 0 && DafnyOptions.O.ArithMode != 3) {
-                return TrToFunctionCall(expr.tok, "Mul", Bpl.Type.Int, oe0, oe1, liftLit);
-              } else {
-                typ = isReal ? Bpl.Type.Real : Bpl.Type.Int;
-                bOpcode = BinaryOperator.Opcode.Mul;
-                break;
-              }
-            case BinaryExpr.ResolvedOpcode.Div:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "div_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.DisableNLarith && !isReal) {
-                return TrToFunctionCall(expr.tok, "INTERNAL_div_boogie", Bpl.Type.Int, e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.ArithMode != 0 && DafnyOptions.O.ArithMode != 3) {
-                return TrToFunctionCall(expr.tok, "Div", Bpl.Type.Int, e0, oe1, liftLit);
-              } else if (isReal) {
-                typ = Bpl.Type.Real;
-                bOpcode = BinaryOperator.Opcode.RealDiv;
-                break;
-              } else {
-                typ = Bpl.Type.Int;
-                bOpcode = BinaryOperator.Opcode.Div;
-                break;
-              }
-            case BinaryExpr.ResolvedOpcode.Mod:
-              if (0 <= bvWidth) {
-                return TrToFunctionCall(expr.tok, "mod_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              } else if (DafnyOptions.O.DisableNLarith && !isReal) {
-                return TrToFunctionCall(expr.tok, "INTERNAL_mod_boogie", Bpl.Type.Int, e0, e1, liftLit);
-              } else if (!isReal && DafnyOptions.O.ArithMode != 0 && DafnyOptions.O.ArithMode != 3) {
-                return TrToFunctionCall(expr.tok, "Mod", Bpl.Type.Int, e0, oe1, liftLit);
-              } else {
-                typ = isReal ? Bpl.Type.Real : Bpl.Type.Int;
-                bOpcode = BinaryOperator.Opcode.Mod;
-                break;
-              }
-
-            case BinaryExpr.ResolvedOpcode.LeftShift: {
-                Contract.Assert(0 <= bvWidth);
-                return TrToFunctionCall(expr.tok, "LeftShift_bv" + bvWidth, translator.BplBvType(bvWidth), e0, translator.ConvertExpression(expr.tok, e1, e.E1.Type, e.Type), liftLit);
-              }
-            case BinaryExpr.ResolvedOpcode.RightShift: {
-                Contract.Assert(0 <= bvWidth);
-                return TrToFunctionCall(expr.tok, "RightShift_bv" + bvWidth, translator.BplBvType(bvWidth), e0, translator.ConvertExpression(expr.tok, e1, e.E1.Type, e.Type), liftLit);
-              }
-            case BinaryExpr.ResolvedOpcode.BitwiseAnd: {
-              Contract.Assert(0 <= bvWidth);
-              return TrToFunctionCall(expr.tok, "and_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-            }
-            case BinaryExpr.ResolvedOpcode.BitwiseOr: {
-                Contract.Assert(0 <= bvWidth);
-                return TrToFunctionCall(expr.tok, "or_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              }
-            case BinaryExpr.ResolvedOpcode.BitwiseXor: {
-                Contract.Assert(0 <= bvWidth);
-                return TrToFunctionCall(expr.tok, "xor_bv" + bvWidth, translator.BplBvType(bvWidth), e0, e1, liftLit);
-              }
-
-            case BinaryExpr.ResolvedOpcode.LtChar:
-            case BinaryExpr.ResolvedOpcode.LeChar:
-            case BinaryExpr.ResolvedOpcode.GeChar:
-            case BinaryExpr.ResolvedOpcode.GtChar: {
-                // work off the original operands (that is, allow them to be lit-wrapped)
-                var operand0 = translator.FunctionCall(e0.tok, BuiltinFunction.CharToInt, null, oe0);
-                var operand1 = translator.FunctionCall(e0.tok, BuiltinFunction.CharToInt, null, oe1);
-                BinaryOperator.Opcode bOp;
-                switch (e.ResolvedOp) {
-                  case BinaryExpr.ResolvedOpcode.LtChar:  bOp = BinaryOperator.Opcode.Lt; break;
-                  case BinaryExpr.ResolvedOpcode.LeChar: bOp = BinaryOperator.Opcode.Le; break;
-                  case BinaryExpr.ResolvedOpcode.GeChar: bOp = BinaryOperator.Opcode.Ge; break;
-                  case BinaryExpr.ResolvedOpcode.GtChar: bOp = BinaryOperator.Opcode.Gt; break;
-                  default:
-                    Contract.Assert(false);  // unexpected case
-                    throw new cce.UnreachableException();  // to please compiler
-                }
-                return Bpl.Expr.Binary(expr.tok, bOp, operand0, operand1);
-              }
-
-            case BinaryExpr.ResolvedOpcode.SetEq:
-            case BinaryExpr.ResolvedOpcode.MultiSetEq:
-            case BinaryExpr.ResolvedOpcode.SeqEq:
-            case BinaryExpr.ResolvedOpcode.MapEq:
-              return translator.TypeSpecificEqual(expr.tok, e.E0.Type, e0, e1);
-            case BinaryExpr.ResolvedOpcode.SetNeq:
-            case BinaryExpr.ResolvedOpcode.MultiSetNeq:
-            case BinaryExpr.ResolvedOpcode.SeqNeq:
-            case BinaryExpr.ResolvedOpcode.MapNeq:
-              return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, translator.TypeSpecificEqual(expr.tok, e.E0.Type, e0, e1));
-
-            case BinaryExpr.ResolvedOpcode.ProperSubset: {
-              return translator.ProperSubset(expr.tok, e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.Subset: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetSubset : BuiltinFunction.ISetSubset;
-              return translator.FunctionCall(expr.tok, f, null, e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.Superset: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetSubset : BuiltinFunction.ISetSubset;
-              return translator.FunctionCall(expr.tok, f, null, e1, e0);
-            }
-            case BinaryExpr.ResolvedOpcode.ProperSuperset:
-              return translator.ProperSubset(expr.tok, e1, e0);
-            case BinaryExpr.ResolvedOpcode.Disjoint: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetDisjoint : BuiltinFunction.ISetDisjoint;
-              return translator.FunctionCall(expr.tok, f, null, e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.InSet:
-              Contract.Assert(false); throw new cce.UnreachableException();  // this case handled above
-            case BinaryExpr.ResolvedOpcode.NotInSet:
-              Contract.Assert(false); throw new cce.UnreachableException();  // this case handled above
-            case BinaryExpr.ResolvedOpcode.Union: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetUnion : BuiltinFunction.ISetUnion;
-              return translator.FunctionCall(expr.tok, f, translator.TrType(expr.Type.AsSetType.Arg), e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.Intersection: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetIntersection : BuiltinFunction.ISetIntersection;
-              return translator.FunctionCall(expr.tok, f, translator.TrType(expr.Type.AsSetType.Arg), e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.SetDifference: {
-              bool finite = e.E1.Type.AsSetType.Finite;
-              var f = finite ? BuiltinFunction.SetDifference : BuiltinFunction.ISetDifference;
-              return translator.FunctionCall(expr.tok, f, translator.TrType(expr.Type.AsSetType.Arg), e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.ProperMultiSubset:
-              return translator.ProperMultiset(expr.tok, e0, e1);
-            case BinaryExpr.ResolvedOpcode.MultiSubset:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetSubset, null, e0, e1);
-            case BinaryExpr.ResolvedOpcode.MultiSuperset:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetSubset, null, e1, e0);
-            case BinaryExpr.ResolvedOpcode.ProperMultiSuperset:
-              return translator.ProperMultiset(expr.tok, e1, e0);
-            case BinaryExpr.ResolvedOpcode.MultiSetDisjoint:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetDisjoint, null, e0, e1);
-            case BinaryExpr.ResolvedOpcode.InMultiSet:
-              Contract.Assert(false); throw new cce.UnreachableException();  // this case handled above
-            case BinaryExpr.ResolvedOpcode.NotInMultiSet:
-              Contract.Assert(false); throw new cce.UnreachableException();  // this case handled above
-            case BinaryExpr.ResolvedOpcode.MultiSetUnion:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetUnion, translator.TrType(expr.Type.AsMultiSetType.Arg), e0, e1);
-            case BinaryExpr.ResolvedOpcode.MultiSetIntersection:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetIntersection, translator.TrType(expr.Type.AsMultiSetType.Arg), e0, e1);
-            case BinaryExpr.ResolvedOpcode.MultiSetDifference:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.MultiSetDifference, translator.TrType(expr.Type.AsMultiSetType.Arg), e0, e1);
-
-            case BinaryExpr.ResolvedOpcode.ProperPrefix:
-              return translator.ProperPrefix(expr.tok, e0, e1);
-            case BinaryExpr.ResolvedOpcode.Prefix:
-              {
-                Bpl.Expr len0 = translator.FunctionCall(expr.tok, BuiltinFunction.SeqLength, null, e0);
-                Bpl.Expr len1 = translator.FunctionCall(expr.tok, BuiltinFunction.SeqLength, null, e1);
-                return Bpl.Expr.Binary(expr.tok, BinaryOperator.Opcode.And,
-                  Bpl.Expr.Le(len0, len1),
-                  translator.FunctionCall(expr.tok, BuiltinFunction.SeqSameUntil, null, e0, e1, len0));
-              }
-            case BinaryExpr.ResolvedOpcode.Concat:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.SeqAppend, translator.TrType(expr.Type.AsSeqType.Arg), e0, e1);
-            case BinaryExpr.ResolvedOpcode.InSeq:
-              return translator.FunctionCall(expr.tok, BuiltinFunction.SeqContains, null, e1,
-                       BoxIfNecessary(expr.tok, e0, cce.NonNull(e.E0.Type)));
-            case BinaryExpr.ResolvedOpcode.NotInSeq:
-              Bpl.Expr arg = translator.FunctionCall(expr.tok, BuiltinFunction.SeqContains, null, e1,
-                       BoxIfNecessary(expr.tok, e0, cce.NonNull(e.E0.Type)));
-              return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, arg);
-            case BinaryExpr.ResolvedOpcode.InMap: {
-              bool finite = e.E1.Type.AsMapType.Finite;
-              var f = finite ? BuiltinFunction.MapDomain : BuiltinFunction.IMapDomain;
-              return Bpl.Expr.SelectTok(expr.tok, translator.FunctionCall(expr.tok, f, predef.MapType(e.tok, finite, predef.BoxType, predef.BoxType), e1),
-                                     BoxIfNecessary(expr.tok, e0, e.E0.Type));
-            }
-            case BinaryExpr.ResolvedOpcode.NotInMap: {
-              bool finite = e.E1.Type.AsMapType.Finite;
-              var f = finite ? BuiltinFunction.MapDomain : BuiltinFunction.IMapDomain;
-              Bpl.Expr inMap = Bpl.Expr.SelectTok(expr.tok, translator.FunctionCall(expr.tok, f, predef.MapType(e.tok, finite, predef.BoxType, predef.BoxType), e1),
-                                     BoxIfNecessary(expr.tok, e0, e.E0.Type));
-              return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, inMap);
-            }
-            case BinaryExpr.ResolvedOpcode.MapMerge: {
-              bool finite = e.E0.Type.AsMapType.Finite;
-              var f = finite ? "Map#Merge" : "IMap#Merge";
-              return translator.FunctionCall(expr.tok, f, translator.TrType(expr.Type), e0, e1);
-            }
-            case BinaryExpr.ResolvedOpcode.MapSubtraction: {
-              bool finite = e.E0.Type.AsMapType.Finite;
-              var f = finite ? "Map#Subtract" : "IMap#Subtract";
-              return translator.FunctionCall(expr.tok, f, translator.TrType(expr.Type), e0, e1);
-            }
-
-            case BinaryExpr.ResolvedOpcode.RankLt:
-              return Bpl.Expr.Binary(expr.tok, BinaryOperator.Opcode.Lt,
-                translator.FunctionCall(expr.tok, e.E0.Type.IsDatatype ? BuiltinFunction.DtRank : BuiltinFunction.BoxRank, null, e0),
-                translator.FunctionCall(expr.tok, BuiltinFunction.DtRank, null, e1));
-            case BinaryExpr.ResolvedOpcode.RankGt:
-              return Bpl.Expr.Binary(expr.tok, BinaryOperator.Opcode.Gt,
-                translator.FunctionCall(expr.tok, BuiltinFunction.DtRank, null, e0),
-                translator.FunctionCall(expr.tok, e.E1.Type.IsDatatype ? BuiltinFunction.DtRank: BuiltinFunction.BoxRank, null, e1));
-
-            default:
-              Contract.Assert(false); throw new cce.UnreachableException();  // unexpected binary expression
-          }
-          liftLit = liftLit && !keepLits;
-          var ae0 = keepLits ? oe0 : e0;
-          var ae1 = keepLits ? oe1 : e1;
-          Bpl.Expr re = Bpl.Expr.Binary(expr.tok, bOpcode, ae0, ae1);
-          if (liftLit) {
-            re = MaybeLit(re, typ);
-          }
-          return re;
-        } else if (expr is TernaryExpr) {
-          var e = (TernaryExpr)expr;
-          var e0 = TrExpr(e.E0);
-          if (!TernaryExpr.PrefixEqUsesNat && !e.E0.Type.IsBigOrdinalType) {
-            e0 = translator.FunctionCall(e0.tok, "ORD#FromNat", predef.BigOrdinalType, e0);
-          }
-          var e1 = TrExpr(e.E1);
-          var e2 = TrExpr(e.E2);
-          switch (e.Op) {
-            case TernaryExpr.Opcode.PrefixEqOp:
-            case TernaryExpr.Opcode.PrefixNeqOp:
-              var e1type = e.E1.Type.NormalizeExpand();
-              var e2type = e.E2.Type.NormalizeExpand();
-              var cot = e1type.AsCoDatatype;
-              Contract.Assert(cot != null);  // the argument types of prefix equality (and prefix disequality) are codatatypes
-              var r = translator.CoEqualCall(cot, e1type.TypeArgs, e2type.TypeArgs, e0, this.layerInterCluster.LayerN((int)FuelSetting.FuelAmount.HIGH), e1, e2);
-              if (e.Op == TernaryExpr.Opcode.PrefixEqOp) {
-                return r;
-              } else {
-                return Bpl.Expr.Unary(expr.tok, UnaryOperator.Opcode.Not, r);
-              }
-            default:
-              Contract.Assert(false); throw new cce.UnreachableException();  // unexpected ternary expression
-          }
-        } else if (expr is LetExpr) {
-          var e = (LetExpr)expr;
-          if (!e.Exact) {
-            var d = translator.LetDesugaring(e);
-            return TrExpr(d);
-          } else {
-            List<Bpl.Variable> lhss;
-            List<Bpl.Expr> rhss;
-            TrLetExprPieces(e, out lhss, out rhss);
-            // in the translation of body, treat a let-bound variable as IsLit if its RHS definition is IsLit
-            Contract.Assert(lhss.Count == rhss.Count);  // this is a postcondition of TrLetExprPieces
-            var previousCount = translator.letBoundVariablesWithLitRHS.Count;
-            for (var i = 0; i < lhss.Count; i++) {
-              if (translator.IsLit(rhss[i])) {
-                translator.letBoundVariablesWithLitRHS.Add(lhss[i].Name);
-              }
-              i++;
-            }
-            var body = TrExpr(e.Body);
-            foreach (var v in lhss) {
-              translator.letBoundVariablesWithLitRHS.Remove(v.Name);
-            }
-            Contract.Assert(previousCount == translator.letBoundVariablesWithLitRHS.Count);
-            // in the following, use the token for Body instead of the token for the whole let expression; this gives better error locations
-            return new Bpl.LetExpr(e.Body.tok, lhss, rhss, null, body);
-          }
-        } else if (expr is QuantifierExpr) {
-          QuantifierExpr e = (QuantifierExpr)expr;
-
-          if (e.SplitQuantifier != null) {
-            return TrExpr(e.SplitQuantifierExpression);
-          } else {
-            List<Variable> tyvars = translator.MkTyParamBinders(e.TypeArgs);
-            List<Variable> bvars = new List<Variable>();
-            var bodyEtran = this;
-            if (e is ExistsExpr && translator.stmtContext == StmtType.ASSERT && translator.adjustFuelForExists) {
-              // assert exists need decrease fuel by 1
-              bodyEtran = bodyEtran.DecreaseFuel(1);
-              // set adjustFuelForExists to false so that we don't keep decrease the fuel in cases like the expr below.
-              // assert exists p:int :: exists t:T :: ToInt(t) > 0;
-              translator.adjustFuelForExists = false;
-            } else if (e is ExistsExpr && translator.stmtContext == StmtType.ASSUME && translator.adjustFuelForExists) {
-              // assume exists need increase fuel by 1
-              bodyEtran = bodyEtran.LayerOffset(1);
-              translator.adjustFuelForExists = false;
-            }
-
-            var initEtran = this;
-            bool _scratch = true;
-
-            Bpl.Expr antecedent = Bpl.Expr.True;
-
-            if (Attributes.ContainsBool(e.Attributes, "layerQuantifier", ref _scratch)) {
-              // If this is a layer quantifier, quantify over layers here, and use $LS(ly) layers in the translation of the body
-              var ly = BplBoundVar(e.Refresh("q$ly#", translator.CurrentIdGenerator), predef.LayerType, bvars);
-              bodyEtran = bodyEtran.ReplaceLayer(ly);
-            }
-            if (Attributes.ContainsBool(e.Attributes, "heapQuantifier", ref _scratch)) {
-              var h = BplBoundVar(e.Refresh("q$heap#", translator.CurrentIdGenerator), predef.HeapType, bvars);
-              bodyEtran = new ExpressionTranslator(bodyEtran, h);
-              antecedent = BplAnd(new List<Bpl.Expr> {
-                antecedent,
-                translator.FunctionCall(e.tok, BuiltinFunction.IsGoodHeap, null, h),
-                translator.HeapSameOrSucc(initEtran.HeapExpr, h)  // initHeapForAllStmt
-              });
-            }
-
-            List<bool> freeOfAlloc = null;
-            if (FrugalHeapUseX) {
-              freeOfAlloc = ComprehensionExpr.BoundedPool.HasBounds(e.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc);
-            }
-            antecedent = BplAnd(antecedent, bodyEtran.TrBoundVariables(e.BoundVars, bvars, false, freeOfAlloc)); // initHeapForAllStmt
-
-            Bpl.QKeyValue kv = TrAttributes(e.Attributes, "trigger");
-            Bpl.Trigger tr = translator.TrTrigger(bodyEtran, e.Attributes, e.tok, bvars, null, null);
-
-            if (e.Range != null) {
-              antecedent = BplAnd(antecedent, bodyEtran.TrExpr(e.Range));
-            }
-            Bpl.Expr body = bodyEtran.TrExpr(e.Term);
-
-            if (e is ForallExpr) {
-              return new Bpl.ForallExpr(expr.tok, new List<TypeVariable>(), Concat(tyvars, bvars), kv, tr, Bpl.Expr.Imp(antecedent, body));
-            } else {
-              Contract.Assert(e is ExistsExpr);
-              return new Bpl.ExistsExpr(expr.tok, new List<TypeVariable>(), Concat(tyvars, bvars), kv, tr, Bpl.Expr.And(antecedent, body));
-            }
-          }
-        } else if (expr is SetComprehension) {
-          var e = (SetComprehension)expr;
-          List<bool> freeOfAlloc = null;
-          if (FrugalHeapUseX) {
-            freeOfAlloc = ComprehensionExpr.BoundedPool.HasBounds(e.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc);
-          }
-          // Translate "set xs | R :: T" into:
-          //     lambda y: BoxType :: (exists xs :: CorrectType(xs) && R && y==Box(T))
-          // or if "T" is "xs", then:
-          //     lambda y: BoxType :: CorrectType(y) && R[xs := Unbox(y)]
-          var yVar = new Bpl.BoundVariable(expr.tok, new Bpl.TypedIdent(expr.tok, translator.CurrentIdGenerator.FreshId("$y#"), predef.BoxType));
-          Bpl.Expr y = new Bpl.IdentifierExpr(expr.tok, yVar);
-          Bpl.Expr lbody;
-          if (e.TermIsSimple) {
-            // lambda y: BoxType :: CorrectType(y) && R[xs := yUnboxed]
-            var bv = e.BoundVars[0];
-            Bpl.Expr typeAntecedent = translator.MkIsBox(new Bpl.IdentifierExpr(expr.tok, yVar), bv.Type);
-            if (freeOfAlloc != null && !freeOfAlloc[0]) {
-              var isAlloc = translator.MkIsAlloc(new Bpl.IdentifierExpr(expr.tok, yVar), bv.Type, HeapExpr);
-              typeAntecedent = BplAnd(typeAntecedent, isAlloc);
-            }
-            var yUnboxed = translator.UnboxIfBoxed(new Bpl.IdentifierExpr(expr.tok, yVar), bv.Type);
-            var range = Translator.Substitute(e.Range, bv, new BoogieWrapper(yUnboxed, bv.Type));
-            lbody = BplAnd(typeAntecedent, TrExpr(range));
-          } else {
-            // lambda y: BoxType :: (exists xs :: CorrectType(xs) && R && y==Box(T))
-            List<Variable> bvars = new List<Variable>();
-            Bpl.Expr typeAntecedent = TrBoundVariables(e.BoundVars, bvars, false, freeOfAlloc);
-
-            var eq = Bpl.Expr.Eq(y, BoxIfNecessary(expr.tok, TrExpr(e.Term), e.Term.Type));
-            var ebody = Bpl.Expr.And(BplAnd(typeAntecedent, TrExpr(e.Range)), eq);
-            var triggers = translator.TrTrigger(this, e.Attributes, e.tok);
-            lbody = new Bpl.ExistsExpr(expr.tok, bvars, triggers, ebody);
-          }
-          Bpl.QKeyValue kv = TrAttributes(e.Attributes, "trigger");
-          return new Bpl.LambdaExpr(expr.tok, new List<TypeVariable>(), new List<Variable> { yVar }, kv, lbody);
-
-        } else if (expr is MapComprehension) {
-          var e = (MapComprehension)expr;
-          // Translate "map x,y | R(x,y) :: F(x,y) := G(x,y)" into
-          // Map#Glue(lambda w: BoxType :: exists x,y :: R(x,y) && unbox(w) == F(x,y),
-          //          lambda w: BoxType :: G(project_x(unbox(w)), project_y(unbox(w))),
-          //          type)".
-          // where project_x and project_y are functions defined (elsewhere, in CanCallAssumption) by the following axiom:
-          //     forall x,y :: R(x,y) ==> var x',y' := project_x(unbox(F(x,y))),project_y(unbox(F(x,y))); R(x',y') && F(x',y') == F(x,y)
-          // that is (without the let expression):
-          //     forall x,y :: R(x,y) ==> R(project_x(unbox(F(x,y))), project_y(unbox(F(x,y)))) && F(project_x(unbox(F(x,y))), project_y(unbox(F(x,y)))) == F(x,y)
-          //
-          // In the common case where F(x,y) is omitted (in which case the list of bound variables is restricted to length 1):
-          // Translate "map x | R(x) :: G(x)" into
-          // Map#Glue(lambda w: BoxType :: R(unbox(w)),
-          //          lambda w: BoxType :: G(unbox(w)),
-          //          type)".
-          List<Variable> bvars = new List<Variable>();
-          List<bool> freeOfAlloc = null;
-          if (FrugalHeapUseX) {
-            freeOfAlloc = ComprehensionExpr.BoundedPool.HasBounds(e.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc);
-          }
-          TrBoundVariables(e.BoundVars, bvars, false, freeOfAlloc);
-
-          Bpl.QKeyValue kv = TrAttributes(e.Attributes, "trigger");
-
-          var wVar = new Bpl.BoundVariable(expr.tok, new Bpl.TypedIdent(expr.tok, translator.CurrentIdGenerator.FreshId("$w#"), predef.BoxType));
-
-          Bpl.Expr keys, values;
-          if (!e.IsGeneralMapComprehension) {
-            var bv = e.BoundVars[0];
-            var w = new Bpl.IdentifierExpr(expr.tok, wVar);
-            Bpl.Expr unboxw = translator.UnboxIfBoxed(w, bv.Type);
-            Bpl.Expr typeAntecedent = translator.MkIsBox(w, bv.Type);
-            var subst = new Dictionary<IVariable,Expression>();
-            subst.Add(bv, new BoogieWrapper(unboxw, bv.Type));
-
-            var ebody = BplAnd(typeAntecedent, TrExpr(Translator.Substitute(e.Range, null, subst)));
-            keys = new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), new List<Variable> { wVar }, kv, ebody);
-            ebody = TrExpr(Translator.Substitute(e.Term, null, subst));
-            values = new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), new List<Variable> { wVar }, kv, BoxIfNecessary(expr.tok, ebody, e.Term.Type));
-          } else {
-            var t = e.TermLeft;
-            var w = new Bpl.IdentifierExpr(expr.tok, wVar);
-            Bpl.Expr unboxw = translator.UnboxIfBoxed(w, t.Type);
-            Bpl.Expr typeAntecedent = translator.MkIsBox(w, t.Type);
-            List<Bpl.Variable> bvs;
-            List<Bpl.Expr> args;
-            translator.CreateBoundVariables(e.BoundVars, out bvs, out args);
-            Contract.Assert(e.BoundVars.Count == bvs.Count);
-            var subst = new Dictionary<IVariable, Expression>();
-            for (var i = 0; i < e.BoundVars.Count; i++) {
-              subst.Add(e.BoundVars[i], new BoogieWrapper(args[i], e.BoundVars[i].Type));
-            }
-            var rr = TrExpr(Translator.Substitute(e.Range, null, subst));
-            var ff = TrExpr(Translator.Substitute(t, null, subst));
-            var exst_body = BplAnd(rr, Bpl.Expr.Eq(unboxw, ff));
-            var ebody = BplAnd(typeAntecedent, new Bpl.ExistsExpr(e.tok, bvs, exst_body));
-            keys = new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), new List<Variable> { wVar }, kv, ebody);
-
-            translator.CreateMapComprehensionProjectionFunctions(e);
-            Contract.Assert(e.ProjectionFunctions != null && e.ProjectionFunctions.Count == e.BoundVars.Count);
-            subst = new Dictionary<IVariable, Expression>();
-            for (var i = 0; i < e.BoundVars.Count; i++) {
-              var p = new Bpl.NAryExpr(e.tok, new Bpl.FunctionCall(e.ProjectionFunctions[i]), new List<Bpl.Expr> { unboxw });
-              var prj = new BoogieWrapper(p, e.BoundVars[i].Type);
-              subst.Add(e.BoundVars[i], prj);
-            }
-            ebody = TrExpr(Translator.Substitute(e.Term, null, subst));
-            values = new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), new List<Variable> { wVar }, kv, BoxIfNecessary(expr.tok, ebody, e.Term.Type));
-          }
-
-          bool finite = e.Finite;
-          var f = finite ? BuiltinFunction.MapGlue : BuiltinFunction.IMapGlue;
-          return translator.FunctionCall(e.tok, f, null, keys, values, translator.TypeToTy(expr.Type));
-
-        } else if (expr is LambdaExpr) {
-          var e = (LambdaExpr)expr;
-          return TrLambdaExpr(e);
-
-        } else if (expr is StmtExpr) {
-          var e = (StmtExpr)expr;
-          return TrExpr(e.E);
-
-        } else if (expr is ITEExpr) {
-          ITEExpr e = (ITEExpr)expr;
-          var g = translator.RemoveLit(TrExpr(e.Test));
-          var thn = translator.RemoveLit(TrExpr(e.Thn));
-          var els = translator.RemoveLit(TrExpr(e.Els));
-          return new NAryExpr(expr.tok, new IfThenElse(expr.tok), new List<Bpl.Expr> { g, thn, els });
-        } else if (expr is MatchExpr) {
-          var e = (MatchExpr)expr;
-          var ite = DesugarMatchExpr(e);
-          return TrExpr(ite);
-
-        } else if (expr is ConcreteSyntaxExpression) {
-          var e = (ConcreteSyntaxExpression)expr;
-          return TrExpr(e.ResolvedExpression);
-
-        } else if (expr is BoxingCastExpr) {
-          BoxingCastExpr e = (BoxingCastExpr)expr;
-          return translator.CondApplyBox(e.tok, TrExpr(e.E), e.FromType, e.ToType);
-
-        } else if (expr is UnboxingCastExpr) {
-          UnboxingCastExpr e = (UnboxingCastExpr)expr;
-          return translator.CondApplyUnbox(e.tok, TrExpr(e.E), e.FromType, e.ToType);
-
-        } else {
-          Contract.Assert(false); throw new cce.UnreachableException();  // unexpected expression
-        }
-      }
-
-      public Expr TrExprSpecialFunctionCall(FunctionCallExpr expr) {
-        Contract.Requires(expr.Function is SpecialFunction);
-        string name = expr.Function.Name;
-        if (name == "RotateLeft") {
-          var w = expr.Type.AsBitVectorType.Width;
-          Expression arg = expr.Args[0];
-          return TrToFunctionCall(expr.tok, "LeftRotate_bv" + w, translator.BplBvType(w), TrExpr(expr.Receiver), translator.ConvertExpression(expr.tok, TrExpr(arg), arg.Type, expr.Type), false);
-        } else if (name == "RotateRight") {
-          var w = expr.Type.AsBitVectorType.Width;
-          Expression arg = expr.Args[0];
-          return TrToFunctionCall(expr.tok, "RightRotate_bv" + w, translator.BplBvType(w), TrExpr(expr.Receiver), translator.ConvertExpression(expr.tok, TrExpr(arg), arg.Type, expr.Type), false);
-        } else {
-          bool argsAreLit_dummy;
-          var args = FunctionInvocationArguments(expr, null, true, out argsAreLit_dummy);
-          var id = new Bpl.IdentifierExpr(expr.tok, expr.Function.FullSanitizedName, translator.TrType(expr.Type));
-          return new Bpl.NAryExpr(expr.tok, new Bpl.FunctionCall(id), args);
-        }
-      }
-      public Expr TrToFunctionCall(IToken tok, string function, Bpl.Type returnType, Bpl.Expr e0, Bpl.Expr e1, bool liftLit) {
-        Bpl.Expr re = translator.FunctionCall(tok, function, returnType, e0, e1);
-        if (liftLit) {
-          re = MaybeLit(re, returnType);
-        }
-        return re;
-      }
-
-      private Expr TrLambdaExpr(LambdaExpr e) {
-        Contract.Requires(e != null);
-
-        var bvars = new List<Bpl.Variable>();
-
-        var varNameGen = translator.CurrentIdGenerator.NestedFreshIdGenerator("$l#");
-
-        var heap = BplBoundVar(varNameGen.FreshId("#heap#"), predef.HeapType, bvars);
-
-        var ves = (from bv in e.BoundVars select
-          BplBoundVar(varNameGen.FreshId(string.Format("#{0}#", bv.Name)), predef.BoxType, bvars)).ToList();
-        var subst = e.BoundVars.Zip(ves, (bv, ve) => {
-          var unboxy = translator.UnboxIfBoxed(ve, bv.Type);
-          return new KeyValuePair<IVariable, Expression>(bv, new BoogieWrapper(unboxy, bv.Type));
-        }).ToDictionary(x => x.Key, x => x.Value);
-        var su = new Substituter(null, subst, new Dictionary<TypeParameter, Type>());
-
-        var et = new ExpressionTranslator(this, heap);
-        var lvars = new List<Bpl.Variable>();
-        var ly = BplBoundVar(varNameGen.FreshId("#ly#"), predef.LayerType, lvars);
-        et = et.WithLayer(ly);
-
-        var ebody = et.TrExpr(Translator.Substitute(e.Body, null, subst));
-        ebody = translator.BoxIfUnboxed(ebody, e.Body.Type);
-
-        var isBoxes = BplAnd(ves.Zip(e.BoundVars, (ve, bv) => translator.MkIsBox(ve, bv.Type)));
-        var reqbody = e.Range == null
-          ? isBoxes
-          : BplAnd(isBoxes, et.TrExpr(Translator.Substitute(e.Range, null, subst)));
-
-        var rdvars = new List<Bpl.Variable>();
-        var o = BplBoundVar(varNameGen.FreshId("#o#"), predef.RefType, rdvars);
-        Bpl.Expr rdbody = new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), rdvars, null,
-          translator.InRWClause(e.tok, o, null, e.Reads.ConvertAll(su.SubstFrameExpr), et, null, null));
-        rdbody = translator.FunctionCall(e.tok, "SetRef_to_SetBox", predef.SetType(e.tok, true, predef.BoxType), rdbody);
-
-        return MaybeLit(
-          translator.FunctionCall(e.tok, BuiltinFunction.AtLayer, predef.HandleType,
-            new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), lvars, null,
-              translator.FunctionCall(e.tok, translator.Handle(e.BoundVars.Count), predef.BoxType,
-                new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), bvars, null, ebody),
-                new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), bvars, null, reqbody),
-                new Bpl.LambdaExpr(e.tok, new List<TypeVariable>(), bvars, null, rdbody))),
-                layerIntraCluster != null ? layerIntraCluster.ToExpr() : layerInterCluster.ToExpr()),
-          predef.HandleType);
-      }
-
-      public void TrLetExprPieces(LetExpr let, out List<Bpl.Variable> lhss, out List<Bpl.Expr> rhss) {
-        Contract.Requires(let != null);
-        var substMap = new Dictionary<IVariable, Expression>();
-        for (int i = 0; i < let.LHSs.Count; i++) {
-          translator.AddCasePatternVarSubstitutions(let.LHSs[i], TrExpr(let.RHSs[i]), substMap);
-        }
-        lhss = new List<Bpl.Variable>();
-        rhss = new List<Bpl.Expr>();
-        foreach (var v in let.BoundVars) {
-          var rhs = substMap[v];  // this should succeed (that is, "v" is in "substMap"), because the AddCasePatternVarSubstitutions calls above should have added a mapping for each bound variable in let.BoundVars
-          Bpl.Expr bvIde;  // unused
-          var bv = BplBoundVar(v.AssignUniqueName(translator.currentDeclaration.IdGenerator), translator.TrType(v.Type), out bvIde);
-          lhss.Add(bv);
-          rhss.Add(TrExpr(rhs));
-        }
-      }
-
-      public Expression DesugarMatchExpr(MatchExpr e) {
-        Contract.Requires(e != null);
-        // Translate:
-        //   match S
-        //   case C(i, j) => X
-        //   case D(k, l) => Y
-        //   case E(m, n) => Z
-        // into:
-        //   if S.C? then
-        //     X[i,j := S.dC0, S.dC1]
-        //   else if S.D? then
-        //     Y[k,l := S.dD0, S.dD1]
-        //   else
-        //     Z[m,n := S.dE0, S.dE1]
-        // As a special case, when there are no cases at all (which, in a correct program, means the
-        // match expression is unreachable), the translation is:
-        //   t
-        // where is "t" is some value (in particular, the default value) of the expected type.
-        Expression r = null;
-        for (int i = e.Cases.Count; 0 <= --i; ) {
-          var mc = e.Cases[i];
-          var substMap = new Dictionary<IVariable, Expression>();
-          var argIndex = 0;
-          foreach (var bv in mc.Arguments) {
-            if (!LocalVariable.HasWildcardName(bv)) {
-              var dtor = mc.Ctor.Destructors[argIndex];
-              var dv = new MemberSelectExpr(bv.tok, e.Source, dtor);
-              substMap.Add(bv, dv);
-            }
-            argIndex++;
-          }
-          var c = Translator.Substitute(mc.Body, null, substMap);
-          if (r == null) {
-            r = c;
-          } else {
-            var test = new MemberSelectExpr(mc.tok, e.Source, mc.Ctor.QueryField);
-            var ite = new ITEExpr(mc.tok, false, test, c, r);
-            ite.Type = e.Type;
-            r = ite;
-          }
-        }
-        return r ?? new BoogieWrapper(ArbitraryValue(e.Type), e.Type);
-      }
-
-      public Bpl.Expr TrBoundVariables(List<BoundVar/*!*/> boundVars, List<Variable> bvars) {
-        return TrBoundVariables(boundVars, bvars, false);
-      }
-
-      public Bpl.Expr TrBoundVariables(List<BoundVar/*!*/> boundVars, List<Variable> bvars, bool translateAsLocals, List<bool>/*?*/ freeOfAlloc = null) {
-        Contract.Requires(boundVars != null);
-        Contract.Requires(bvars != null);
-        Contract.Requires(freeOfAlloc == null || freeOfAlloc.Count == boundVars.Count);
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-        Bpl.Expr typeAntecedent = Bpl.Expr.True;
-        var i = 0;
-        foreach (BoundVar bv in boundVars) {
-          var tid = new Bpl.TypedIdent(bv.tok, bv.AssignUniqueName(translator.currentDeclaration.IdGenerator), translator.TrType(bv.Type));
-          Bpl.Variable bvar;
-          if (translateAsLocals) {
-            bvar = new Bpl.LocalVariable(bv.tok, tid);
-          } else {
-            bvar = new Bpl.BoundVariable(bv.tok, tid);
-          }
-          bvars.Add(bvar);
-          var useAlloc = freeOfAlloc == null || freeOfAlloc[i] ? NOALLOC : ISALLOC;
-          Bpl.Expr wh = translator.GetWhereClause(bv.tok, new Bpl.IdentifierExpr(bv.tok, bvar), bv.Type, this, useAlloc);
-          if (wh != null) {
-            typeAntecedent = BplAnd(typeAntecedent, wh);
-          }
-          i++;
-        }
-        return typeAntecedent;
-      }
-
-      public List<Tuple<Bpl.Variable, Bpl.Expr>> TrBoundVariables_SeparateWhereClauses(List<BoundVar/*!*/> boundVars) {
-        Contract.Requires(boundVars != null);
-        Contract.Ensures(Contract.Result<List<Tuple<Bpl.Variable, Bpl.Expr>>>() != null);
-
-        var varsAndAntecedents = new List<Tuple<Bpl.Variable, Bpl.Expr>>();
-        foreach (BoundVar bv in boundVars) {
-          var tid = new Bpl.TypedIdent(bv.tok, bv.AssignUniqueName(translator.currentDeclaration.IdGenerator), translator.TrType(bv.Type));
-          var bvar = new Bpl.BoundVariable(bv.tok, tid);
-          var wh = translator.GetWhereClause(bv.tok, new Bpl.IdentifierExpr(bv.tok, bvar), bv.Type, this, NOALLOC);
-          varsAndAntecedents.Add(Tuple.Create<Bpl.Variable, Bpl.Expr>(bvar, wh));
-        }
-        return varsAndAntecedents;
-      }
-
-      public Bpl.Expr TrBoundVariablesRename(List<BoundVar> boundVars, List<Variable> bvars, out Dictionary<IVariable, Expression> substMap, out Bpl.Trigger antitriggers) {
-        Contract.Requires(boundVars != null);
-        Contract.Requires(bvars != null);
-
-        substMap = new Dictionary<IVariable, Expression>();
-        antitriggers = null;
-        Bpl.Expr typeAntecedent = Bpl.Expr.True;
-        foreach (BoundVar bv in boundVars) {
-          var newBoundVar = new BoundVar(bv.tok, bv.Name, bv.Type);
-          IdentifierExpr ie = new IdentifierExpr(newBoundVar.tok, newBoundVar.AssignUniqueName(translator.currentDeclaration.IdGenerator));
-          ie.Var = newBoundVar; ie.Type = ie.Var.Type;  // resolve ie here
-          substMap.Add(bv, ie);
-          Bpl.Variable bvar = new Bpl.BoundVariable(newBoundVar.tok, new Bpl.TypedIdent(newBoundVar.tok, newBoundVar.AssignUniqueName(translator.currentDeclaration.IdGenerator), translator.TrType(newBoundVar.Type)));
-          bvars.Add(bvar);
-          var bIe = new Bpl.IdentifierExpr(bvar.tok, bvar);
-          Bpl.Expr wh = translator.GetWhereClause(bv.tok, bIe, newBoundVar.Type, this, NOALLOC);
-          if (wh != null) {
-            typeAntecedent = BplAnd(typeAntecedent, wh);
-          }
-        }
-        return typeAntecedent;
-      }
-
-      public List<Bpl.Expr> FunctionInvocationArguments(FunctionCallExpr e, Bpl.Expr layerArgument) {
-        bool dummy;
-        return FunctionInvocationArguments(e, layerArgument, false, out dummy);
-      }
-
-      public List<Bpl.Expr> FunctionInvocationArguments(FunctionCallExpr e, Bpl.Expr layerArgument, bool omitHeapArgument, out bool argsAreLit) {
-        Contract.Requires(e != null);
-        Contract.Ensures(Contract.Result<List<Bpl.Expr>>() != null);
-
-        var args = new List<Bpl.Expr>();
-
-        // first add type arguments
-        var tyParams = GetTypeParams(e.Function);
-        var tySubst = e.TypeArgumentSubstitutionsWithParents();
-        args.AddRange(translator.trTypeArgs(tySubst, tyParams));
-
-        if (layerArgument != null) {
-          args.Add(layerArgument);
-        }
-        if (e.Function is TwoStateFunction) {
-          args.Add(OldAt(e.AtLabel).HeapExpr);
-        }
-        if (!omitHeapArgument && (AlwaysUseHeap || e.Function.ReadsHeap)) {
-          Contract.Assert(HeapExpr != null);
-          args.Add(HeapExpr);
-          // if the function doesn't use heaps, but always use heap as an argument
-          // we want to quantify over the heap so that heap in the trigger can match over
-          // heap modifying operations. (see Dafny4/bug144.dfy)
-          bool useHeap = e.Function.ReadsHeap;
-          if (!useHeap) {
-            foreach (var arg in e.Function.Formals) {
-              if (arg.Type.IsRefType) {
-                useHeap = true;
-                break;
-              }
-            }
-          }
-          if (!useHeap) {
-            Statistics_HeapAsQuantifierCount++;
-          }
-        }
-        argsAreLit = true;
-        if (!e.Function.IsStatic) {
-          var tr_ee = TrExpr(e.Receiver);
-          argsAreLit = argsAreLit && translator.IsLit(tr_ee);
-          args.Add(tr_ee);
-        }
-        for (int i = 0; i < e.Args.Count; i++) {
-          Expression ee = e.Args[i];
-          Type t = e.Function.Formals[i].Type;
-          Expr tr_ee = TrExpr(ee);
-          argsAreLit = argsAreLit && translator.IsLit(tr_ee);
-          args.Add(translator.CondApplyBox(e.tok, tr_ee, cce.NonNull(ee.Type), t));
-        }
-        return args;
-      }
-
-      public Bpl.Expr GetArrayIndexFieldName(IToken tok, List<Expression> indices) {
-        return translator.GetArrayIndexFieldName(tok, indices.ConvertAll(idx => {
-          var e = TrExpr(idx);
-          return translator.ConvertExpression(idx.tok, e, idx.Type, Type.Int);
-        }));
-      }
-
-      public Bpl.Expr BoxIfNecessary(IToken tok, Bpl.Expr e, Type fromType) {
-        Contract.Requires(tok != null);
-        Contract.Requires(e != null);
-        Contract.Requires(fromType != null);
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-        return translator.BoxIfNecessary(tok, e, fromType);
-      }
-
-      public static Bpl.NAryExpr ReadHeap(IToken tok, Expr heap, Expr r, Expr f) {
-        Contract.Requires(tok != null);
-        Contract.Requires(heap != null);
-        Contract.Requires(r != null);
-        Contract.Requires(f != null);
-        Contract.Ensures(Contract.Result<Bpl.NAryExpr>() != null);
-
-        List<Bpl.Expr> args = new List<Bpl.Expr>();
-        args.Add(heap);
-        args.Add(r);
-        args.Add(f);
-        Bpl.Type t = (f.Type != null) ? f.Type : f.ShallowType;
-        return new Bpl.NAryExpr(tok,
-          new Bpl.FunctionCall(new Bpl.IdentifierExpr(tok, "read", t.AsCtor.Arguments[0])),
-          args);
-      }
-
-
-      public static Bpl.NAryExpr UpdateHeap(IToken tok, Expr heap, Expr r, Expr f, Expr v) {
-        Contract.Requires(tok != null);
-        Contract.Requires(heap != null);
-        Contract.Requires(r != null);
-        Contract.Requires(f != null);
-        Contract.Requires(v != null);
-        Contract.Ensures(Contract.Result<Bpl.NAryExpr>() != null);
-
-        List<Bpl.Expr> args = new List<Bpl.Expr>();
-        args.Add(heap);
-        args.Add(r);
-        args.Add(f);
-        args.Add(v);
-        return new Bpl.NAryExpr(tok,
-          new Bpl.FunctionCall(new Bpl.IdentifierExpr(tok, "update", heap.Type)),
-          args);
-      }
-
-      /// <summary>
-      /// Translate like s[Box(elmt)], but try to avoid as many set functions as possible in the
-      /// translation, because such functions can mess up triggering.
-      /// </summary>
-      public Bpl.Expr TrInSet(IToken tok, Bpl.Expr elmt, Expression s, Type elmtType, bool aggressive, out bool performedRewrite) {
-        Contract.Requires(tok != null);
-        Contract.Requires(elmt != null);
-        Contract.Requires(s != null);
-        Contract.Requires(elmtType != null);
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-        var elmtBox = BoxIfNecessary(tok, elmt, elmtType);
-        var r = TrInSet_Aux(tok, elmt, elmtBox, s, aggressive, out performedRewrite);
-        Contract.Assert(performedRewrite == RewriteInExpr(s, aggressive)); // sanity check
-        return r;
-      }
-      /// <summary>
-      /// The worker routine for TrInSet.  This method takes both "elmt" and "elmtBox" as parameters,
-      /// using the former when the unboxed form is needed and the latter when the boxed form is needed.
-      /// This gives the caller the flexibility to pass in either "o, Box(o)" or "Unbox(bx), bx".
-      /// Note: This method must be kept in synch with RewriteInExpr.
-      /// </summary>
-      public Bpl.Expr TrInSet_Aux(IToken tok, Bpl.Expr elmt, Bpl.Expr elmtBox, Expression s, bool aggressive, out bool performedRewrite) {
-        Contract.Requires(tok != null);
-        Contract.Requires(elmt != null);
-        Contract.Requires(elmtBox != null);
-        Contract.Requires(s != null);
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-        performedRewrite = true;  // assume a rewrite will happen
-        s = s.Resolved;
-        bool pr;
-        if (s is BinaryExpr && aggressive) {
-          BinaryExpr bin = (BinaryExpr)s;
-          switch (bin.ResolvedOp) {
-            case BinaryExpr.ResolvedOpcode.Union:
-              return Bpl.Expr.Or(TrInSet_Aux(tok, elmt, elmtBox, bin.E0, aggressive, out pr), TrInSet_Aux(tok, elmt, elmtBox, bin.E1, aggressive, out pr));
-            case BinaryExpr.ResolvedOpcode.Intersection:
-              return Bpl.Expr.And(TrInSet_Aux(tok, elmt, elmtBox, bin.E0, aggressive, out pr), TrInSet_Aux(tok, elmt, elmtBox, bin.E1, aggressive, out pr));
-            case BinaryExpr.ResolvedOpcode.SetDifference:
-              return Bpl.Expr.And(TrInSet_Aux(tok, elmt, elmtBox, bin.E0, aggressive, out pr), Bpl.Expr.Not(TrInSet_Aux(tok, elmt, elmtBox, bin.E1, aggressive, out pr)));
-            default:
-              break;
-          }
-        } else if (s is SetDisplayExpr) {
-          SetDisplayExpr disp = (SetDisplayExpr)s;
-          Bpl.Expr disjunction = null;
-          foreach (Expression a in disp.Elements) {
-            Bpl.Expr disjunct = Bpl.Expr.Eq(elmt, TrExpr(a));
-            if (disjunction == null) {
-              disjunction = disjunct;
-            } else {
-              disjunction = Bpl.Expr.Or(disjunction, disjunct);
-            }
-          }
-          if (disjunction == null) {
-            return Bpl.Expr.False;
-          } else {
-            return disjunction;
-          }
-        } else if (s is SetComprehension) {
-          var compr = (SetComprehension)s;
-          // Translate "elmt in set xs | R :: T" into:
-          //     exists xs :: CorrectType(xs) && R && elmt==T
-          // or if "T" is "xs", then:
-          //     CorrectType(elmt) && R[xs := elmt]
-          if (compr.TermIsSimple) {
-            // CorrectType(elmt) && R[xs := elmt]
-            // Note, we can always use NOALLOC here.
-            Bpl.Expr typeAntecedent = translator.GetWhereClause(compr.tok, elmt, compr.BoundVars[0].Type, this, NOALLOC) ?? Bpl.Expr.True;
-            var range = Translator.Substitute(compr.Range, compr.BoundVars[0], new BoogieWrapper(elmt, compr.BoundVars[0].Type));
-            return BplAnd(typeAntecedent, TrExpr(range));
-          } else {
-            // exists xs :: CorrectType(xs) && R && elmt==T
-            List<bool> freeOfAlloc = null;
-            if (FrugalHeapUseX) {
-              freeOfAlloc = ComprehensionExpr.BoundedPool.HasBounds(compr.Bounds, ComprehensionExpr.BoundedPool.PoolVirtues.IndependentOfAlloc_or_ExplicitAlloc);
-            }
-            var bvars = new List<Variable>();
-            Bpl.Expr typeAntecedent = TrBoundVariables(compr.BoundVars, bvars, false, freeOfAlloc) ?? Bpl.Expr.True;
-            var eq = Bpl.Expr.Eq(elmtBox, BoxIfNecessary(compr.tok, TrExpr(compr.Term), compr.Term.Type));
-            var ebody = Bpl.Expr.And(BplAnd(typeAntecedent, TrExpr(compr.Range)), eq);
-            var triggers = translator.TrTrigger(this, compr.Attributes, compr.tok);
-            return new Bpl.ExistsExpr(compr.tok, bvars, triggers, ebody);
-          }
-        }
-        performedRewrite = false;
-        return Bpl.Expr.SelectTok(tok, TrExpr(s), elmtBox);
-      }
-
-      /// <summary>
-      /// Translate like 0 < s[Box(elmt)], but try to avoid as many set functions as possible in the
-      /// translation, because such functions can mess up triggering.
-      /// Note: This method must be kept in synch with RewriteInExpr.
-      /// </summary>
-      public Bpl.Expr TrInMultiSet(IToken tok, Bpl.Expr elmt, Expression s, Type elmtType, bool aggressive) {
-        Contract.Requires(tok != null);
-        Contract.Requires(elmt != null);
-        Contract.Requires(s != null);
-        Contract.Requires(elmtType != null);
-
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-        var elmtBox = BoxIfNecessary(tok, elmt, elmtType);
-        return TrInMultiSet_Aux(tok, elmt, elmtBox, s, aggressive);
-      }
-      public Bpl.Expr TrInMultiSet_Aux(IToken tok, Bpl.Expr elmt, Bpl.Expr elmtBox, Expression s, bool aggressive) {
-        Contract.Requires(tok != null);
-        Contract.Requires(elmt != null);
-        Contract.Requires(s != null);
-        Contract.Requires(elmtBox != null);
-
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-        s = s.Resolved;
-        if (s is BinaryExpr && aggressive) {
-          BinaryExpr bin = (BinaryExpr)s;
-          switch (bin.ResolvedOp) {
-            case BinaryExpr.ResolvedOpcode.MultiSetUnion:
-              return Bpl.Expr.Binary(tok, BinaryOperator.Opcode.Or, TrInMultiSet_Aux(tok, elmt, elmtBox, bin.E0, aggressive), TrInMultiSet_Aux(tok, elmt, elmtBox, bin.E1, aggressive));
-            case BinaryExpr.ResolvedOpcode.MultiSetIntersection:
-              return Bpl.Expr.Binary(tok, BinaryOperator.Opcode.And, TrInMultiSet_Aux(tok, elmt, elmtBox, bin.E0, aggressive), TrInMultiSet_Aux(tok, elmt, elmtBox, bin.E1, aggressive));
-            default:
-              break;
-          }
-        } else if (s is MultiSetDisplayExpr) {
-          MultiSetDisplayExpr disp = (MultiSetDisplayExpr)s;
-          Bpl.Expr disjunction = null;
-          foreach (Expression a in disp.Elements) {
-            Bpl.Expr disjunct = Bpl.Expr.Eq(elmt, TrExpr(a));
-            if (disjunction == null) {
-              disjunction = disjunct;
-            } else {
-              disjunction = Bpl.Expr.Or(disjunction, disjunct);
-            }
-          }
-          if (disjunction == null) {
-            return Bpl.Expr.False;
-          } else {
-            return disjunction;
-          }
-        }
-        return Bpl.Expr.Gt(Bpl.Expr.SelectTok(tok, TrExpr(s), elmtBox), Bpl.Expr.Literal(0));
-      }
-
-      /// <summary>
-      /// This method returns "true" iff TrInSet_Aux/TrInMultiSet_Aux will rewrite an expression "x in s".
-      /// Note: This method must be kept in synch with TrInSet_Aux/TrInMultiSet_Aux.
-      /// </summary>
-      public static bool RewriteInExpr(Expression s, bool aggressive) {
-        Contract.Requires(s != null);
-
-        s = s.Resolved;
-        if (s is BinaryExpr && aggressive) {
-          BinaryExpr bin = (BinaryExpr)s;
-          switch (bin.ResolvedOp) {
-            case BinaryExpr.ResolvedOpcode.Union:
-            case BinaryExpr.ResolvedOpcode.Intersection:
-            case BinaryExpr.ResolvedOpcode.SetDifference:
-            case BinaryExpr.ResolvedOpcode.MultiSetUnion:
-            case BinaryExpr.ResolvedOpcode.MultiSetIntersection:
-              return true;
-            default:
-              break;
-          }
-        } else if (s is SetDisplayExpr || s is MultiSetDisplayExpr) {
-          return true;
-        } else if (s is SetComprehension) {
-          return true;
-        }
-        return false;
-      }
-
-      public Bpl.QKeyValue TrAttributes(Attributes attrs, string skipThisAttribute) {
-        Bpl.QKeyValue kv = null;
-        bool hasNewTimeLimit = Attributes.Contains(attrs, "_timeLimit");
-        bool hasNewRLimit = Attributes.Contains(attrs, "_rlimit");
-        foreach (var attr in attrs.AsEnumerable()) {
-          if (attr.Name == skipThisAttribute
-           || attr.Name == "axiom"  // Dafny's axiom attribute clashes with Boogie's axiom keyword
-           || attr.Name == "fuel"   // Fuel often uses function names as arguments, which adds extra axioms unnecessarily
-           || (DafnyOptions.O.DisallowExterns && (attr.Name == "extern" || attr.Name == "dllimport")) // omit the extern attribute when /noExterns option is specified.
-           || attr.Name == "timeLimitMultiplier"  // This is a Dafny-specific attribute
-           || (attr.Name == "timeLimit" && hasNewTimeLimit)
-           || (attr.Name == "rlimit" && hasNewRLimit)
-             ) {
-            continue;
-          }
-          List<object> parms = new List<object>();
-          foreach (var arg in attr.Args) {
-            var s = arg.AsStringLiteral();
-            if (s != null) {
-              // pass string literals down to Boogie as string literals, not as their expression translation
-              parms.Add(s);
-            } else {
-              var e = TrExpr(arg);
-              e = translator.RemoveLit(e);
-              parms.Add(e);
-            }
-          }
-
-          var name = attr.Name;
-          if (name == "_timeLimit") {
-            name = "timeLimit";
-          } else if (name == "_rlimit") {
-            name = "rlimit";
-          }
-          kv = new Bpl.QKeyValue(Token.NoToken, name, parms, kv);
-        }
-        return kv;
-      }
-
-      // --------------- help routines ---------------
-
-      public Bpl.Expr IsAlloced(IToken tok, Bpl.Expr e) {
-        Contract.Requires(HeapExpr != null);
-        return translator.IsAlloced(tok, HeapExpr, e);
-      }
-
-      public Bpl.Expr GoodRef(IToken tok, Bpl.Expr e, Type type) {
-        Contract.Requires(tok != null);
-        Contract.Requires(e != null);
-        Contract.Requires(type != null);
-        Contract.Ensures(Contract.Result<Bpl.Expr>() != null);
-
-        // Add $Is and $IsAlloc
-        return translator.GetWhereClause(tok, e, type, this, ISALLOC);
       }
     }
